% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcquisitionSessionsApi.r
\name{AcquisitionSessionsApi}
\alias{AcquisitionSessionsApi}
\title{AcquisitionSessions operations}
\description{
phisWSClientRTools.AcquisitionSessions
}
\section{Methods}{

\describe{

get4_p_metadata_file Get the metadata for the acquisition session file

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AcquisitionSessionsApi$new()}}
\item \href{#method-get4_p_metadata_file}{\code{AcquisitionSessionsApi$get4_p_metadata_file()}}
\item \href{#method-clone}{\code{AcquisitionSessionsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcquisitionSessionsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get4_p_metadata_file"></a>}}
\subsection{Method \code{get4_p_metadata_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcquisitionSessionsApi$get4_p_metadata_file(
  vector_rdf_type,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcquisitionSessionsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
