% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActuatorDetailDTO.r
\docType{data}
\name{ActuatorDetailDTO}
\alias{ActuatorDetailDTO}
\title{ActuatorDetailDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ActuatorDetailDTO
}
\description{
ActuatorDetailDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{model}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{dateOfLastCalibration}}{}

\item{\code{personInCharge}}{}

\item{\code{variables}}{}
}}

\keyword{datasets}
