% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActuatorsApi.r
\name{ActuatorsApi}
\alias{ActuatorsApi}
\title{Actuators operations}
\description{
phisWSClientRTools.Actuators
}
\section{Methods}{

\describe{

get_actuator_details Get an actuator


get_actuators_by_search Get all actuators corresponding to the search params given


post Post actuator(s)


put Put actuator(s)


put_measured_variables Update the measured variables of an actuator

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ActuatorsApi$new()}}
\item \href{#method-get_actuator_details}{\code{ActuatorsApi$get_actuator_details()}}
\item \href{#method-get_actuators_by_search}{\code{ActuatorsApi$get_actuators_by_search()}}
\item \href{#method-post}{\code{ActuatorsApi$post()}}
\item \href{#method-put}{\code{ActuatorsApi$put()}}
\item \href{#method-put_measured_variables}{\code{ActuatorsApi$put_measured_variables()}}
\item \href{#method-clone}{\code{ActuatorsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_actuator_details"></a>}}
\subsection{Method \code{get_actuator_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorsApi$get_actuator_details(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_actuators_by_search"></a>}}
\subsection{Method \code{get_actuators_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorsApi$get_actuators_by_search(
  page_size,
  page,
  uri,
  rdf_type,
  label,
  brand,
  serial_number,
  model,
  in_service_date,
  date_of_purchase,
  date_of_last_calibration,
  person_in_charge,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post"></a>}}
\subsection{Method \code{post()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorsApi$post(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put"></a>}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorsApi$put(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_measured_variables"></a>}}
\subsection{Method \code{put_measured_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorsApi$put_measured_variables(uri, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
