% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgronomicalData.r
\docType{data}
\name{AgronomicalData}
\alias{AgronomicalData}
\title{AgronomicalData Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AgronomicalData
}
\description{
AgronomicalData Class
}
\section{Fields}{

\describe{
\item{\code{agronomicalObject}}{}

\item{\code{date}}{}

\item{\code{value}}{}

\item{\code{variable}}{}

\item{\code{sensor}}{}

\item{\code{incertitude}}{}
}}

\keyword{datasets}
