% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgronomicalDataDTO.r
\docType{data}
\name{AgronomicalDataDTO}
\alias{AgronomicalDataDTO}
\title{AgronomicalDataDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AgronomicalDataDTO
}
\description{
AgronomicalDataDTO Class
}
\section{Fields}{

\describe{
\item{\code{agronomicalObject}}{}

\item{\code{date}}{}

\item{\code{value}}{}

\item{\code{sensor}}{}

\item{\code{incertitude}}{}
}}

\keyword{datasets}
