% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationDTO.r
\docType{data}
\name{AnnotationDTO}
\alias{AnnotationDTO}
\title{AnnotationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationDTO
}
\description{
AnnotationDTO Class
}
\section{Fields}{

\describe{
\item{\code{creator}}{}

\item{\code{motivatedBy}}{}

\item{\code{bodyValues}}{}

\item{\code{targets}}{}

\item{\code{uri}}{}

\item{\code{creationDate}}{}
}}

\keyword{datasets}
