% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPostDTO.r
\docType{data}
\name{AnnotationPostDTO}
\alias{AnnotationPostDTO}
\title{AnnotationPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationPostDTO
}
\description{
AnnotationPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{creator}}{}

\item{\code{motivatedBy}}{}

\item{\code{bodyValues}}{}

\item{\code{targets}}{}
}}

\keyword{datasets}
