% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationsApi.r
\docType{data}
\name{AnnotationsApi}
\alias{AnnotationsApi}
\title{Annotations operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationsApi
}
\description{
opensilexWSClient.Annotations
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

delete_annotation_by_uri Delete a list of annotation


get_annotation_by_uri Get a annotation


get_annotations_by_search Get all annotations corresponding to the search params given


post1 Post annotations

}
}

\keyword{datasets}
