% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiDescriptionDTO.r
\name{ApiDescriptionDTO}
\alias{ApiDescriptionDTO}
\title{ApiDescriptionDTO Class}
\description{
ApiDescriptionDTO Class

ApiDescriptionDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{buildVersion}}{}

\item{\code{majorVersion}}{}

\item{\code{name}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ApiDescriptionDTO$new()}}
\item \href{#method-toJSON}{\code{ApiDescriptionDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ApiDescriptionDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ApiDescriptionDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ApiDescriptionDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ApiDescriptionDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ApiDescriptionDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiDescriptionDTO$new(buildVersion, majorVersion, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiDescriptionDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiDescriptionDTO$fromJSON(ApiDescriptionDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiDescriptionDTO$fromJSONObject(ApiDescriptionDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiDescriptionDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiDescriptionDTO$fromJSONString(ApiDescriptionDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApiDescriptionDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
