% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiGermplasmDTO.r
\docType{data}
\name{BrapiGermplasmDTO}
\alias{BrapiGermplasmDTO}
\title{BrapiGermplasmDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiGermplasmDTO
}
\description{
BrapiGermplasmDTO Class
}
\section{Fields}{

\describe{
\item{\code{accessionNumber}}{}

\item{\code{acquisitionDate}}{}

\item{\code{biologicalStatusOfAccessionCode}}{}

\item{\code{breedingMethodDbId}}{}

\item{\code{commonCropName}}{}

\item{\code{countryOfOrigin}}{}

\item{\code{defaultDisplayName}}{}

\item{\code{documentationURL}}{}

\item{\code{donors}}{}

\item{\code{genus}}{}

\item{\code{germplasmGenus}}{}

\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{germplasmPUI}}{}

\item{\code{germplasmSpecies}}{}

\item{\code{instituteCode}}{}

\item{\code{instituteName}}{}

\item{\code{pedigree}}{}

\item{\code{seedsource}}{}

\item{\code{species}}{}

\item{\code{speciesAuthority}}{}

\item{\code{subtaxa}}{}

\item{\code{subtaxaAuthority}}{}

\item{\code{synonyms}}{}

\item{\code{taxonIds}}{}

\item{\code{typeOfGermplasmStorageCode}}{}
}}

\keyword{datasets}
