% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiMethod.r
\name{BrapiMethod}
\alias{BrapiMethod}
\title{BrapiMethod Class}
\description{
BrapiMethod Class

BrapiMethod Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{methodDbId}}{}

\item{\code{methodName}}{}

\item{\code{brapiClass}}{}

\item{\code{description}}{}

\item{\code{formula}}{}

\item{\code{ontologyReference}}{}

\item{\code{reference}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiMethod$new()}}
\item \href{#method-toJSON}{\code{BrapiMethod$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiMethod$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiMethod$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiMethod$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiMethod$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiMethod$new(
  methodDbId,
  methodName,
  brapiClass,
  description,
  formula,
  ontologyReference,
  reference
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiMethod$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiMethod$fromJSON(BrapiMethodJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiMethod$fromJSONObject(BrapiMethodObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiMethod$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiMethod$fromJSONString(BrapiMethodJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
