% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiObservationDTO.r
\name{BrapiObservationDTO}
\alias{BrapiObservationDTO}
\title{BrapiObservationDTO Class}
\description{
BrapiObservationDTO Class

BrapiObservationDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{observationDbId}}{}

\item{\code{observationLevel}}{}

\item{\code{observationTimeStamp}}{}

\item{\code{observationUnitDbId}}{}

\item{\code{observationUnitName}}{}

\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{operator}}{}

\item{\code{season}}{}

\item{\code{studyDbId}}{}

\item{\code{uploadedBy}}{}

\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiObservationDTO$new()}}
\item \href{#method-toJSON}{\code{BrapiObservationDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiObservationDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiObservationDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiObservationDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiObservationDTO$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiObservationDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationDTO$new(
  germplasmDbId,
  germplasmName,
  observationDbId,
  observationLevel,
  observationTimeStamp,
  observationUnitDbId,
  observationUnitName,
  observationVariableDbId,
  observationVariableName,
  operator,
  season,
  studyDbId,
  uploadedBy,
  value
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationDTO$fromJSON(BrapiObservationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationDTO$fromJSONObject(BrapiObservationDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationDTO$fromJSONString(BrapiObservationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
