% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiObservationSummaryDTO.r
\docType{data}
\name{BrapiObservationSummaryDTO}
\alias{BrapiObservationSummaryDTO}
\title{BrapiObservationSummaryDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiObservationSummaryDTO
}
\description{
BrapiObservationSummaryDTO Class
}
\section{Fields}{

\describe{
\item{\code{collector}}{}

\item{\code{observationDbId}}{}

\item{\code{observationTimeStamp}}{}

\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{season}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
