% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiObservationUnitDTO.r
\docType{data}
\name{BrapiObservationUnitDTO}
\alias{BrapiObservationUnitDTO}
\title{BrapiObservationUnitDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiObservationUnitDTO
}
\description{
BrapiObservationUnitDTO Class
}
\section{Fields}{

\describe{
\item{\code{blockNumber}}{}

\item{\code{entryNumber}}{}

\item{\code{entryType}}{}

\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{locationDbId}}{}

\item{\code{locationName}}{}

\item{\code{observationLevel}}{}

\item{\code{observationLevels}}{}

\item{\code{observationUnitDbId}}{}

\item{\code{observationUnitName}}{}

\item{\code{observationUnitXref}}{}

\item{\code{observations}}{}

\item{\code{pedigree}}{}

\item{\code{plantNumber}}{}

\item{\code{plotNumber}}{}

\item{\code{positionCoordinateX}}{}

\item{\code{positionCoordinateXType}}{}

\item{\code{positionCoordinateY}}{}

\item{\code{positionCoordinateYType}}{}

\item{\code{programDbId}}{}

\item{\code{programName}}{}

\item{\code{replicate}}{}

\item{\code{studyDbId}}{}

\item{\code{studyName}}{}

\item{\code{treatments}}{}

\item{\code{trialDbId}}{}

\item{\code{trialName}}{}
}}

\keyword{datasets}
