% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiScale.r
\name{BrapiScale}
\alias{BrapiScale}
\title{BrapiScale Class}
\description{
BrapiScale Class

BrapiScale Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scaleDbid}}{}

\item{\code{scaleName}}{}

\item{\code{dataType}}{}

\item{\code{decimalPlaces}}{}

\item{\code{ontologyReference}}{}

\item{\code{xref}}{}

\item{\code{validValues}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiScale$new()}}
\item \href{#method-toJSON}{\code{BrapiScale$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiScale$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiScale$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiScale$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiScale$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiScale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiScale$new(
  scaleDbid,
  scaleName,
  dataType,
  decimalPlaces,
  ontologyReference,
  xref,
  validValues
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiScale$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiScale$fromJSON(BrapiScaleJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiScale$fromJSONObject(BrapiScaleObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiScale$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiScale$fromJSONString(BrapiScaleJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiScale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
