% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiTraitDTO.r
\docType{data}
\name{BrapiTraitDTO}
\alias{BrapiTraitDTO}
\title{BrapiTraitDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiTraitDTO
}
\description{
BrapiTraitDTO Class
}
\section{Fields}{

\describe{
\item{\code{defaultValue}}{}

\item{\code{description}}{}

\item{\code{traitName}}{}

\item{\code{observationVariables}}{}

\item{\code{traitId}}{}

\item{\code{traitDbId}}{}
}}

\keyword{datasets}
