% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1callsApi.r
\name{Brapiv1callsApi}
\alias{Brapiv1callsApi}
\title{Brapiv1calls operations}
\description{
phisWSClientRTools.Brapiv1calls
}
\section{Methods}{

\describe{

get_calls Check the available BrAPI calls

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Brapiv1callsApi$new()}}
\item \href{#method-get_calls}{\code{Brapiv1callsApi$get_calls()}}
\item \href{#method-clone}{\code{Brapiv1callsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1callsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_calls"></a>}}
\subsection{Method \code{get_calls()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1callsApi$get_calls(page_size, page, data_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1callsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
