% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1germplasmApi.r
\docType{data}
\name{Brapiv1germplasmApi}
\alias{Brapiv1germplasmApi}
\title{Brapiv1germplasm operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Brapiv1germplasmApi
}
\description{
opensilexWSClient.Brapiv1germplasm
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_germplasm_by_search1 Get all germplasm corresponding to the search params given

}
}

\keyword{datasets}
