% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1studiesApi.r
\name{Brapiv1studiesApi}
\alias{Brapiv1studiesApi}
\title{Brapiv1studies operations}
\description{
phisWSClientRTools.Brapiv1studies
}
\section{Methods}{

\describe{

get_observation_units List all the observation units measured in the study.


get_observation_variables List all the observation variables measured in the study.


get_observations Get the observations associated to a specific study


get_studies Retrieve studies information


get_study_details Retrieve study details

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Brapiv1studiesApi$new()}}
\item \href{#method-get_observation_units}{\code{Brapiv1studiesApi$get_observation_units()}}
\item \href{#method-get_observation_variables}{\code{Brapiv1studiesApi$get_observation_variables()}}
\item \href{#method-get_observations}{\code{Brapiv1studiesApi$get_observations()}}
\item \href{#method-get_studies}{\code{Brapiv1studiesApi$get_studies()}}
\item \href{#method-get_study_details}{\code{Brapiv1studiesApi$get_study_details()}}
\item \href{#method-clone}{\code{Brapiv1studiesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_observation_units"></a>}}
\subsection{Method \code{get_observation_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesApi$get_observation_units(
  study_db_id,
  observation_level,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_observation_variables"></a>}}
\subsection{Method \code{get_observation_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesApi$get_observation_variables(study_db_id, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_observations"></a>}}
\subsection{Method \code{get_observations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesApi$get_observations(
  study_db_id,
  observation_variable_db_ids,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_studies"></a>}}
\subsection{Method \code{get_studies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesApi$get_studies(
  study_db_id,
  common_crop_name,
  season_db_id,
  active,
  sort_by,
  sort_order,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_study_details"></a>}}
\subsection{Method \code{get_study_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesApi$get_study_details(study_db_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
