% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1studiesSearchApi.r
\docType{data}
\name{Brapiv1studiesSearchApi}
\alias{Brapiv1studiesSearchApi}
\title{Brapiv1studiesSearch operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Brapiv1studiesSearchApi
}
\description{
opensilexWSClient.Brapiv1studiesSearch
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_studies_search Retrieve studies information


post_studies_search search studies

}
}

\keyword{datasets}
