% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1tokenApi.r
\name{Brapiv1tokenApi}
\alias{Brapiv1tokenApi}
\title{Brapiv1token operations}
\description{
phisWSClientRTools.Brapiv1token
}
\section{Methods}{

\describe{

get_token Login


log_out Log out

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Brapiv1tokenApi$new()}}
\item \href{#method-get_token}{\code{Brapiv1tokenApi$get_token()}}
\item \href{#method-log_out}{\code{Brapiv1tokenApi$log_out()}}
\item \href{#method-clone}{\code{Brapiv1tokenApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1tokenApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_token"></a>}}
\subsection{Method \code{get_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1tokenApi$get_token(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_out"></a>}}
\subsection{Method \code{log_out()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1tokenApi$log_out(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1tokenApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
