% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1traitsApi.r
\name{Brapiv1traitsApi}
\alias{Brapiv1traitsApi}
\title{Brapiv1traits operations}
\description{
phisWSClientRTools.Brapiv1traits
}
\section{Methods}{

\describe{

get_trait_details1 Retrieve trait details by id


get_traits_list Retrieve the list of all traits available in the system

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Brapiv1traitsApi$new()}}
\item \href{#method-get_trait_details1}{\code{Brapiv1traitsApi$get_trait_details1()}}
\item \href{#method-get_traits_list}{\code{Brapiv1traitsApi$get_traits_list()}}
\item \href{#method-clone}{\code{Brapiv1traitsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1traitsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_trait_details1"></a>}}
\subsection{Method \code{get_trait_details1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1traitsApi$get_trait_details1(trait_db_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_traits_list"></a>}}
\subsection{Method \code{get_traits_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1traitsApi$get_traits_list(page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1traitsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
