% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1variablesApi.r
\name{Brapiv1variablesApi}
\alias{Brapiv1variablesApi}
\title{Brapiv1variables operations}
\description{
phisWSClientRTools.Brapiv1variables
}
\section{Methods}{

\describe{

get_observation_variables1 List all the observation variables measured in the study.


get_variable_details Retrieve variable details by id


get_variables_list Retrieve the list of all variables available in the system

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Brapiv1variablesApi$new()}}
\item \href{#method-get_observation_variables1}{\code{Brapiv1variablesApi$get_observation_variables1()}}
\item \href{#method-get_variable_details}{\code{Brapiv1variablesApi$get_variable_details()}}
\item \href{#method-get_variables_list}{\code{Brapiv1variablesApi$get_variables_list()}}
\item \href{#method-clone}{\code{Brapiv1variablesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1variablesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_observation_variables1"></a>}}
\subsection{Method \code{get_observation_variables1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1variablesApi$get_observation_variables1(
  study_db_id,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable_details"></a>}}
\subsection{Method \code{get_variable_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1variablesApi$get_variable_details(observation_variable_db_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variables_list"></a>}}
\subsection{Method \code{get_variables_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1variablesApi$get_variables_list(page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1variablesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
