% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiVariable.r
\docType{data}
\name{BrapiVariable}
\alias{BrapiVariable}
\title{BrapiVariable Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiVariable
}
\description{
BrapiVariable Class
}
\section{Fields}{

\describe{
\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{ontologyReference}}{}

\item{\code{synonyms}}{}

\item{\code{contextOfUse}}{}

\item{\code{growthStage}}{}

\item{\code{status}}{}

\item{\code{xref}}{}

\item{\code{institution}}{}

\item{\code{scientist}}{}

\item{\code{submissionTimesTamp}}{}

\item{\code{language}}{}

\item{\code{crop}}{}

\item{\code{trait}}{}

\item{\code{method}}{}

\item{\code{scale}}{}

\item{\code{defaultValue}}{}

\item{\code{documentationURL}}{}
}}

\keyword{datasets}
