% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiVariableTrait.r
\name{BrapiVariableTrait}
\alias{BrapiVariableTrait}
\title{BrapiVariableTrait Class}
\description{
BrapiVariableTrait Class

BrapiVariableTrait Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{traitDbId}}{}

\item{\code{traitName}}{}

\item{\code{brapiclass}}{}

\item{\code{description}}{}

\item{\code{synonyms}}{}

\item{\code{mainAbbreviation}}{}

\item{\code{alternativeAbbreviations}}{}

\item{\code{entity}}{}

\item{\code{attribute}}{}

\item{\code{status}}{}

\item{\code{xref}}{}

\item{\code{ontologyReference}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiVariableTrait$new()}}
\item \href{#method-toJSON}{\code{BrapiVariableTrait$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiVariableTrait$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiVariableTrait$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiVariableTrait$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiVariableTrait$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiVariableTrait$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariableTrait$new(
  traitDbId,
  traitName,
  brapiclass,
  description,
  synonyms,
  mainAbbreviation,
  alternativeAbbreviations,
  entity,
  attribute,
  status,
  xref,
  ontologyReference
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariableTrait$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariableTrait$fromJSON(BrapiVariableTraitJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariableTrait$fromJSONObject(BrapiVariableTraitObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariableTrait$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariableTrait$fromJSONString(BrapiVariableTraitJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariableTrait$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
