% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcernedItemDTO.r
\docType{data}
\name{ConcernedItemDTO}
\alias{ConcernedItemDTO}
\title{ConcernedItemDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConcernedItemDTO
}
\description{
ConcernedItemDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{typeURI}}{}
}}

\keyword{datasets}
