% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcernedItemWithLabelsDTO.r
\docType{data}
\name{ConcernedItemWithLabelsDTO}
\alias{ConcernedItemWithLabelsDTO}
\title{ConcernedItemWithLabelsDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConcernedItemWithLabelsDTO
}
\description{
ConcernedItemWithLabelsDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{typeURI}}{}

\item{\code{labels}}{}
}}

\keyword{datasets}
