% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contact.r
\docType{data}
\name{Contact}
\alias{Contact}
\title{Contact Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Contact
}
\description{
Contact Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{email}}{}

\item{\code{firstname}}{}

\item{\code{lastname}}{}
}}

\keyword{datasets}
