% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactBrapi.r
\name{ContactBrapi}
\alias{ContactBrapi}
\title{ContactBrapi Class}
\description{
ContactBrapi Class

ContactBrapi Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contactDbId}}{}

\item{\code{name}}{}

\item{\code{instituteName}}{}

\item{\code{email}}{}

\item{\code{type}}{}

\item{\code{orcid}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContactBrapi$new()}}
\item \href{#method-toJSON}{\code{ContactBrapi$toJSON()}}
\item \href{#method-fromJSON}{\code{ContactBrapi$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ContactBrapi$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ContactBrapi$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ContactBrapi$fromJSONString()}}
\item \href{#method-clone}{\code{ContactBrapi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactBrapi$new(contactDbId, name, instituteName, email, type, orcid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactBrapi$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactBrapi$fromJSON(ContactBrapiJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactBrapi$fromJSONObject(ContactBrapiObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactBrapi$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactBrapi$fromJSONString(ContactBrapiJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactBrapi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
