% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactPostgreSQL.r
\docType{data}
\name{ContactPostgreSQL}
\alias{ContactPostgreSQL}
\title{ContactPostgreSQL Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContactPostgreSQL
}
\description{
ContactPostgreSQL Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{firstName}}{}

\item{\code{familyName}}{}

\item{\code{type}}{}
}}

\keyword{datasets}
