% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAnalysisApi.r
\docType{data}
\name{DataAnalysisApi}
\alias{DataAnalysisApi}
\title{DataAnalysis operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataAnalysisApi
}
\description{
opensilexWSClient.DataAnalysis
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_rfunction_results Get a json value


shiny_proxy_server_app_list Get a server availables apps


shiny_proxy_server_status Get a server status message

}
}

\keyword{datasets}
