% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataApi.r
\name{DataApi}
\alias{DataApi}
\title{Data operations}
\description{
phisWSClientRTools.Data
}
\section{Methods}{

\describe{

get_data Get data corresponding to the search parameters given.


get_data_file Get data file


get_data_file_description Get data file description


get_data_file_descriptions_by_search Retrieve data file descriptions corresponding to the search parameters given.


get_data_search Get data corresponding to the search parameters given.


post_data Post data


post_data_file Post data file


post_data_file_path Post data file

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataApi$new()}}
\item \href{#method-get_data}{\code{DataApi$get_data()}}
\item \href{#method-get_data_file}{\code{DataApi$get_data_file()}}
\item \href{#method-get_data_file_description}{\code{DataApi$get_data_file_description()}}
\item \href{#method-get_data_file_descriptions_by_search}{\code{DataApi$get_data_file_descriptions_by_search()}}
\item \href{#method-get_data_search}{\code{DataApi$get_data_search()}}
\item \href{#method-post_data}{\code{DataApi$post_data()}}
\item \href{#method-post_data_file}{\code{DataApi$post_data_file()}}
\item \href{#method-post_data_file_path}{\code{DataApi$post_data_file_path()}}
\item \href{#method-clone}{\code{DataApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data(
  variable,
  page_size,
  page,
  start_date,
  end_date,
  object,
  provenance,
  date_sort_asc,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_file"></a>}}
\subsection{Method \code{get_data_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data_file(file_uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_file_description"></a>}}
\subsection{Method \code{get_data_file_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data_file_description(file_uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_file_descriptions_by_search"></a>}}
\subsection{Method \code{get_data_file_descriptions_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data_file_descriptions_by_search(
  rdf_type,
  page_size,
  page,
  start_date,
  end_date,
  provenance,
  concerned_items,
  json_value_filter,
  date_sort_asc,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_search"></a>}}
\subsection{Method \code{get_data_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$get_data_search(
  page_size,
  page,
  variable_uri,
  start_date,
  end_date,
  object_uri,
  object_label,
  provenance_uri,
  provenance_label,
  date_sort_asc,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_data"></a>}}
\subsection{Method \code{post_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$post_data(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_data_file"></a>}}
\subsection{Method \code{post_data_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$post_data_file(description, file, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_data_file_path"></a>}}
\subsection{Method \code{post_data_file_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$post_data_file_path(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
