% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQueryLog.r
\docType{data}
\name{DataQueryLog}
\alias{DataQueryLog}
\title{DataQueryLog Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataQueryLog
}
\description{
DataQueryLog Class
}
\section{Fields}{

\describe{
\item{\code{userUri}}{}

\item{\code{query}}{}

\item{\code{date}}{}

\item{\code{remoteAdress}}{}
}}

\keyword{datasets}
