% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDTO.r
\docType{data}
\name{DatasetDTO}
\alias{DatasetDTO}
\title{DatasetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DatasetDTO
}
\description{
DatasetDTO Class
}
\section{Fields}{

\describe{
\item{\code{variableUri}}{}

\item{\code{provenance}}{}

\item{\code{data}}{}
}}

\keyword{datasets}
