% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsApi.r
\name{DatasetsApi}
\alias{DatasetsApi}
\title{Datasets operations}
\description{
phisWSClientRTools.Datasets
}
\section{Methods}{

\describe{

get_data_by_search Get all data corresponding to the search params given


post_dataset_data Post dataset

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetsApi$new()}}
\item \href{#method-get_data_by_search}{\code{DatasetsApi$get_data_by_search()}}
\item \href{#method-post_dataset_data}{\code{DatasetsApi$post_dataset_data()}}
\item \href{#method-clone}{\code{DatasetsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_by_search"></a>}}
\subsection{Method \code{get_data_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetsApi$get_data_by_search(
  page_size,
  page,
  experiment,
  variable,
  agronomical_objects,
  start_date,
  end_date,
  sensor,
  incertitude,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_dataset_data"></a>}}
\subsection{Method \code{post_dataset_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetsApi$post_dataset_data(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
