% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteDTO.r
\docType{data}
\name{DeleteDTO}
\alias{DeleteDTO}
\title{DeleteDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DeleteDTO
}
\description{
DeleteDTO Class
}
\section{Fields}{

\describe{
\item{\code{uris}}{}
}}

\keyword{datasets}
