% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentMetadataDTO.r
\docType{data}
\name{DocumentMetadataDTO}
\alias{DocumentMetadataDTO}
\title{DocumentMetadataDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentMetadataDTO
}
\description{
DocumentMetadataDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{documentType}}{}

\item{\code{checksum}}{}

\item{\code{creator}}{}

\item{\code{language}}{}

\item{\code{title}}{}

\item{\code{creationDate}}{}

\item{\code{extension}}{}

\item{\code{comment}}{}

\item{\code{concernedItems}}{}

\item{\code{status}}{}

\item{\code{serverFilePath}}{}
}}

\keyword{datasets}
