% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentsApi.r
\name{DocumentsApi}
\alias{DocumentsApi}
\title{Documents operations}
\description{
phisWSClientRTools.Documents
}
\section{Methods}{

\describe{

get_document_by_uri Get a document (by receiving it&#39;s uri)


get_documents_metadata_by_search Get all documents metadata corresponding to the searched params given


get_documents_type Get all documents types


post_document_file Post data file


post_documents Save a file


put_document_metadata Update document metadata

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DocumentsApi$new()}}
\item \href{#method-get_document_by_uri}{\code{DocumentsApi$get_document_by_uri()}}
\item \href{#method-get_documents_metadata_by_search}{\code{DocumentsApi$get_documents_metadata_by_search()}}
\item \href{#method-get_documents_type}{\code{DocumentsApi$get_documents_type()}}
\item \href{#method-post_document_file}{\code{DocumentsApi$post_document_file()}}
\item \href{#method-post_documents}{\code{DocumentsApi$post_documents()}}
\item \href{#method-put_document_metadata}{\code{DocumentsApi$put_document_metadata()}}
\item \href{#method-clone}{\code{DocumentsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_document_by_uri"></a>}}
\subsection{Method \code{get_document_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$get_document_by_uri(document_uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_documents_metadata_by_search"></a>}}
\subsection{Method \code{get_documents_metadata_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$get_documents_metadata_by_search(
  page_size,
  page,
  uri,
  document_type,
  creator,
  language,
  title,
  creation_date,
  extension,
  concerned_item,
  status,
  sort_by_date,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_documents_type"></a>}}
\subsection{Method \code{get_documents_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$get_documents_type(page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_document_file"></a>}}
\subsection{Method \code{post_document_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$post_document_file(body, uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_documents"></a>}}
\subsection{Method \code{post_documents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$post_documents(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_document_metadata"></a>}}
\subsection{Method \code{put_document_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$put_document_metadata(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
