% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentMeasureDTO.r
\name{EnvironmentMeasureDTO}
\alias{EnvironmentMeasureDTO}
\title{EnvironmentMeasureDTO Class}
\description{
EnvironmentMeasureDTO Class

EnvironmentMeasureDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sensorUri}}{}

\item{\code{variableUri}}{}

\item{\code{date}}{}

\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EnvironmentMeasureDTO$new()}}
\item \href{#method-toJSON}{\code{EnvironmentMeasureDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{EnvironmentMeasureDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{EnvironmentMeasureDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{EnvironmentMeasureDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{EnvironmentMeasureDTO$fromJSONString()}}
\item \href{#method-clone}{\code{EnvironmentMeasureDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasureDTO$new(sensorUri, variableUri, date, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasureDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasureDTO$fromJSON(EnvironmentMeasureDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasureDTO$fromJSONObject(EnvironmentMeasureDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasureDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasureDTO$fromJSONString(EnvironmentMeasureDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasureDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
