% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentMeasurePostDTO.r
\docType{data}
\name{EnvironmentMeasurePostDTO}
\alias{EnvironmentMeasurePostDTO}
\title{EnvironmentMeasurePostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EnvironmentMeasurePostDTO
}
\description{
EnvironmentMeasurePostDTO Class
}
\section{Fields}{

\describe{
\item{\code{sensorUri}}{}

\item{\code{variableUri}}{}

\item{\code{date}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
