% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentsApi.r
\name{EnvironmentsApi}
\alias{EnvironmentsApi}
\title{Environments operations}
\description{
phisWSClientRTools.Environments
}
\section{Methods}{

\describe{

get_environment_measures Get all environment measures corresponding to the search params given


post_environment_measures Post environment(s) measures

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EnvironmentsApi$new()}}
\item \href{#method-get_environment_measures}{\code{EnvironmentsApi$get_environment_measures()}}
\item \href{#method-post_environment_measures}{\code{EnvironmentsApi$post_environment_measures()}}
\item \href{#method-clone}{\code{EnvironmentsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_environment_measures"></a>}}
\subsection{Method \code{get_environment_measures()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentsApi$get_environment_measures(
  variable,
  page_size,
  page,
  start_date,
  end_date,
  sensor,
  date_sort_asc,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_environment_measures"></a>}}
\subsection{Method \code{post_environment_measures()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentsApi$post_environment_measures(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
