% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventGetDTO.r
\docType{data}
\name{EventGetDTO}
\alias{EventGetDTO}
\title{EventGetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventGetDTO
}
\description{
EventGetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{concernedItems}}{}

\item{\code{date}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
