% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventsApi.r
\docType{data}
\name{EventsApi}
\alias{EventsApi}
\title{Events operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventsApi
}
\description{
phisWSClientRTools.Events
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_event_annotations Get an event&#39;s annotations


get_event_by_uri Get the event corresponding to the search uri


get_events Get all events corresponding to the search parameters given.


post2 POST event(s)


put1 Update events

}
}

\keyword{datasets}
