% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.r
\docType{data}
\name{Experiment}
\alias{Experiment}
\title{Experiment Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Experiment
}
\description{
Experiment Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{field}}{}

\item{\code{campaign}}{}

\item{\code{place}}{}

\item{\code{alias}}{}

\item{\code{comment}}{}

\item{\code{keywords}}{}

\item{\code{objective}}{}

\item{\code{cropSpecies}}{}

\item{\code{projects}}{}

\item{\code{groups}}{}

\item{\code{contacts}}{}

\item{\code{variables}}{}

\item{\code{sensors}}{}
}}

\keyword{datasets}
