% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentPostDTO.r
\docType{data}
\name{ExperimentPostDTO}
\alias{ExperimentPostDTO}
\title{ExperimentPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExperimentPostDTO
}
\description{
ExperimentPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{field}}{}

\item{\code{campaign}}{}

\item{\code{place}}{}

\item{\code{alias}}{}

\item{\code{comment}}{}

\item{\code{keywords}}{}

\item{\code{objective}}{}

\item{\code{cropSpecies}}{}

\item{\code{projectsUris}}{}

\item{\code{groupsUris}}{}

\item{\code{contacts}}{}
}}

\keyword{datasets}
