% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentsApi.r
\name{ExperimentsApi}
\alias{ExperimentsApi}
\title{Experiments operations}
\description{
phisWSClientRTools.Experiments
}
\section{Methods}{

\describe{

get_experiment_detail Get an experiment


get_experiments_by_search Get all experiments corresponding to the searched params given


post_experiment Post a experiment


put_experiment Update experiment


put_sensors Update the sensors which participates in an experiment


put_variables Update the observed variables of an experiment

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExperimentsApi$new()}}
\item \href{#method-get_experiment_detail}{\code{ExperimentsApi$get_experiment_detail()}}
\item \href{#method-get_experiments_by_search}{\code{ExperimentsApi$get_experiments_by_search()}}
\item \href{#method-post_experiment}{\code{ExperimentsApi$post_experiment()}}
\item \href{#method-put_experiment}{\code{ExperimentsApi$put_experiment()}}
\item \href{#method-put_sensors}{\code{ExperimentsApi$put_sensors()}}
\item \href{#method-put_variables}{\code{ExperimentsApi$put_variables()}}
\item \href{#method-clone}{\code{ExperimentsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiment_detail"></a>}}
\subsection{Method \code{get_experiment_detail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_experiment_detail(experiment, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiments_by_search"></a>}}
\subsection{Method \code{get_experiments_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$get_experiments_by_search(
  page_size,
  page,
  uri,
  project_uri,
  start_date,
  end_date,
  field,
  campaign,
  place,
  alias,
  keywords,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_experiment"></a>}}
\subsection{Method \code{post_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$post_experiment(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_experiment"></a>}}
\subsection{Method \code{put_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$put_experiment(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_sensors"></a>}}
\subsection{Method \code{put_sensors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$put_sensors(uri, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_variables"></a>}}
\subsection{Method \code{put_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$put_variables(uri, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
