% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorDTO.r
\docType{data}
\name{FactorDTO}
\alias{FactorDTO}
\title{FactorDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FactorDTO
}
\description{
FactorDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}
}}

\keyword{datasets}
