% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileDescriptionDTO.r
\name{FileDescriptionDTO}
\alias{FileDescriptionDTO}
\title{FileDescriptionDTO Class}
\description{
FileDescriptionDTO Class

FileDescriptionDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{date}}{}

\item{\code{concernedItems}}{}

\item{\code{provenanceUri}}{}

\item{\code{metadata}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FileDescriptionDTO$new()}}
\item \href{#method-toJSON}{\code{FileDescriptionDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{FileDescriptionDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{FileDescriptionDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{FileDescriptionDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FileDescriptionDTO$fromJSONString()}}
\item \href{#method-clone}{\code{FileDescriptionDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionDTO$new(
  uri,
  rdfType,
  date,
  concernedItems,
  provenanceUri,
  metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionDTO$fromJSON(FileDescriptionDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionDTO$fromJSONObject(FileDescriptionDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionDTO$fromJSONString(FileDescriptionDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
