% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileDescriptionWebPathPostDTO.r
\docType{data}
\name{FileDescriptionWebPathPostDTO}
\alias{FileDescriptionWebPathPostDTO}
\title{FileDescriptionWebPathPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileDescriptionWebPathPostDTO
}
\description{
FileDescriptionWebPathPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{date}}{}

\item{\code{webPath}}{}

\item{\code{concernedItems}}{}

\item{\code{provenanceUri}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
