% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInformationDTO.r
\docType{data}
\name{FileInformationDTO}
\alias{FileInformationDTO}
\title{FileInformationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileInformationDTO
}
\description{
FileInformationDTO Class
}
\section{Fields}{

\describe{
\item{\code{checksum}}{}

\item{\code{extension}}{}
}}

\keyword{datasets}
