% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileInformations.r
\docType{data}
\name{FileInformations}
\alias{FileInformations}
\title{FileInformations Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileInformations
}
\description{
FileInformations Class
}
\section{Fields}{

\describe{
\item{\code{checksum}}{}

\item{\code{extension}}{}

\item{\code{serverFilePath}}{}
}}

\keyword{datasets}
