% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmApi.r
\docType{data}
\name{GermplasmApi}
\alias{GermplasmApi}
\title{Germplasm operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GermplasmApi
}
\description{
opensilexWSClient.Germplasm
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_germplasm_by_search Get all germplasm corresponding to the search params given


post3 Post Germplasm,

}
}

\keyword{datasets}
