% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermplasmPostDTO.r
\docType{data}
\name{GermplasmPostDTO}
\alias{GermplasmPostDTO}
\title{GermplasmPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GermplasmPostDTO
}
\description{
GermplasmPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{properties}}{}

\item{\code{uri}}{}
}}

\keyword{datasets}
