% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupApi.r
\docType{data}
\name{GroupApi}
\alias{GroupApi}
\title{Group operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupApi
}
\description{
phisWSClientRTools.Group
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_group_by_search Get all groups corresponding to the searched params given


get_group_details Get a group


post_group Post a group


put_group Update groups

}
}

\keyword{datasets}
