% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupPostDTO.r
\docType{data}
\name{GroupPostDTO}
\alias{GroupPostDTO}
\title{GroupPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupPostDTO
}
\description{
GroupPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{level}}{}

\item{\code{description}}{}

\item{\code{usersEmails}}{}
}}

\keyword{datasets}
