% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageMetadata.r
\docType{data}
\name{ImageMetadata}
\alias{ImageMetadata}
\title{ImageMetadata Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ImageMetadata
}
\description{
ImageMetadata Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{concernedItems}}{}

\item{\code{configuration}}{}

\item{\code{fileInformations}}{}
}}

\keyword{datasets}
