% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageMetadataDTO.r
\name{ImageMetadataDTO}
\alias{ImageMetadataDTO}
\title{ImageMetadataDTO Class}
\description{
ImageMetadataDTO Class

ImageMetadataDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rdfType}}{}

\item{\code{concernedItems}}{}

\item{\code{configuration}}{}

\item{\code{fileInfo}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImageMetadataDTO$new()}}
\item \href{#method-toJSON}{\code{ImageMetadataDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ImageMetadataDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ImageMetadataDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ImageMetadataDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ImageMetadataDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ImageMetadataDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataDTO$new(rdfType, concernedItems, configuration, fileInfo)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataDTO$fromJSON(ImageMetadataDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataDTO$fromJSONObject(ImageMetadataDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataDTO$fromJSONString(ImageMetadataDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageMetadataDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
