% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImagesApi.r
\name{ImagesApi}
\alias{ImagesApi}
\title{Images operations}
\description{
phisWSClientRTools.Images
}
\section{Methods}{

\describe{

get_images_by_search Get all images corresponding to the search params given


post_image_file Post data file


post_images_metadata Save a file

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImagesApi$new()}}
\item \href{#method-get_images_by_search}{\code{ImagesApi$get_images_by_search()}}
\item \href{#method-post_image_file}{\code{ImagesApi$post_image_file()}}
\item \href{#method-post_images_metadata}{\code{ImagesApi$post_images_metadata()}}
\item \href{#method-clone}{\code{ImagesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_images_by_search"></a>}}
\subsection{Method \code{get_images_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesApi$get_images_by_search(
  page_size,
  page,
  uri,
  rdf_type,
  concerned_items,
  start_date,
  end_date,
  sensor,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_image_file"></a>}}
\subsection{Method \code{post_image_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesApi$post_image_file(body, uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_images_metadata"></a>}}
\subsection{Method \code{post_images_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesApi$post_images_metadata(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
