% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructuresApi.r
\name{InfrastructuresApi}
\alias{InfrastructuresApi}
\title{Infrastructures operations}
\description{
phisWSClientRTools.Infrastructures
}
\section{Methods}{

\describe{

get_infrastructure_details Get all infrastructure&#39;s details corresponding to the search uri


get_infrastructures_by_search Get all infrastructures corresponding to the search params given

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InfrastructuresApi$new()}}
\item \href{#method-get_infrastructure_details}{\code{InfrastructuresApi$get_infrastructure_details()}}
\item \href{#method-get_infrastructures_by_search}{\code{InfrastructuresApi$get_infrastructures_by_search()}}
\item \href{#method-clone}{\code{InfrastructuresApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructuresApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_infrastructure_details"></a>}}
\subsection{Method \code{get_infrastructure_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructuresApi$get_infrastructure_details(
  uri,
  language,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_infrastructures_by_search"></a>}}
\subsection{Method \code{get_infrastructures_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructuresApi$get_infrastructures_by_search(
  page_size,
  page,
  uri,
  rdf_type,
  label,
  language,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InfrastructuresApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
