% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsonArray.r
\docType{data}
\name{JsonArray}
\alias{JsonArray}
\title{JsonArray Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JsonArray
}
\description{
JsonArray Class
}
\section{Fields}{

\describe{
\item{\code{asString}}{}

\item{\code{asNumber}}{}

\item{\code{asFloat}}{}

\item{\code{asByte}}{}

\item{\code{asCharacter}}{}

\item{\code{asBigDecimal}}{}

\item{\code{asBigInteger}}{}

\item{\code{asShort}}{}

\item{\code{asBoolean}}{}

\item{\code{asInt}}{}

\item{\code{asDouble}}{}

\item{\code{asLong}}{}

\item{\code{asJsonArray}}{}

\item{\code{asJsonPrimitive}}{}

\item{\code{asJsonObject}}{}

\item{\code{jsonObject}}{}

\item{\code{jsonPrimitive}}{}

\item{\code{jsonNull}}{}

\item{\code{jsonArray}}{}

\item{\code{asJsonNull}}{}
}}

\keyword{datasets}
