% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerDTO.r
\docType{data}
\name{LayerDTO}
\alias{LayerDTO}
\title{LayerDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LayerDTO
}
\description{
LayerDTO Class
}
\section{Fields}{

\describe{
\item{\code{objectUri}}{}

\item{\code{objectType}}{}

\item{\code{depth}}{}

\item{\code{generateFile}}{}
}}

\keyword{datasets}
