% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogApi.r
\docType{data}
\name{LogApi}
\alias{LogApi}
\title{Log operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LogApi
}
\description{
phisWSClientRTools.Log
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_data_query_log_search Get data query logs

}
}

\keyword{datasets}
