% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.r
\docType{data}
\name{Metadata}
\alias{Metadata}
\title{Metadata Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Metadata
}
\description{
Metadata Class
}
\section{Fields}{

\describe{
\item{\code{pagination}}{}

\item{\code{status}}{}

\item{\code{datafiles}}{}
}}

\keyword{datasets}
