% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataFileUAVDTO.r
\docType{data}
\name{MetadataFileUAVDTO}
\alias{MetadataFileUAVDTO}
\title{MetadataFileUAVDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MetadataFileUAVDTO
}
\description{
MetadataFileUAVDTO Class
}
\section{Fields}{

\describe{
\item{\code{installation}}{}

\item{\code{groupPlotType}}{}

\item{\code{groupPlotAlias}}{}

\item{\code{groupPlotUri}}{}

\item{\code{groupPlotSpecies}}{}

\item{\code{pilot}}{}

\item{\code{cameraType}}{}

\item{\code{cameraAlias}}{}

\item{\code{cameraUri}}{}

\item{\code{vectorType}}{}

\item{\code{vectorAlias}}{}

\item{\code{vectorUri}}{}

\item{\code{radiometricTargetAlias}}{}

\item{\code{radiometricTargetUri}}{}
}}

\keyword{datasets}
