% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodDTO.r
\docType{data}
\name{MethodDTO}
\alias{MethodDTO}
\title{MethodDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MethodDTO
}
\description{
MethodDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}
}}

\keyword{datasets}
