% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsApi.r
\name{MethodsApi}
\alias{MethodsApi}
\title{Methods operations}
\description{
phisWSClientRTools.Methods
}
\section{Methods}{

\describe{

get_method_details Get a method


get_methods_by_search Get all methods corresponding to the searched params given


post_method Post method(s)


put_method Update method

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MethodsApi$new()}}
\item \href{#method-get_method_details}{\code{MethodsApi$get_method_details()}}
\item \href{#method-get_methods_by_search}{\code{MethodsApi$get_methods_by_search()}}
\item \href{#method-post_method}{\code{MethodsApi$post_method()}}
\item \href{#method-put_method}{\code{MethodsApi$put_method()}}
\item \href{#method-clone}{\code{MethodsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethodsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_method_details"></a>}}
\subsection{Method \code{get_method_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethodsApi$get_method_details(method, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_methods_by_search"></a>}}
\subsection{Method \code{get_methods_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethodsApi$get_methods_by_search(page_size, page, uri, label, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_method"></a>}}
\subsection{Method \code{post_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethodsApi$post_method(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_method"></a>}}
\subsection{Method \code{put_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethodsApi$put_method(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MethodsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
