% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyReference.r
\docType{data}
\name{OntologyReference}
\alias{OntologyReference}
\title{OntologyReference Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OntologyReference
}
\description{
OntologyReference Class
}
\section{Fields}{

\describe{
\item{\code{property}}{}

\item{\code{object}}{}

\item{\code{seeAlso}}{}
}}

\keyword{datasets}
