% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pagination.r
\docType{data}
\name{Pagination}
\alias{Pagination}
\title{Pagination Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Pagination
}
\description{
Pagination Class
}
\section{Fields}{

\describe{
\item{\code{pageSize}}{}

\item{\code{currentPage}}{}

\item{\code{totalCount}}{}

\item{\code{totalPages}}{}
}}

\keyword{datasets}
