% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectDTO.r
\docType{data}
\name{ProjectDTO}
\alias{ProjectDTO}
\title{ProjectDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectDTO
}
\description{
ProjectDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{financialFunding}}{}

\item{\code{financialReference}}{}

\item{\code{description}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{homePage}}{}

\item{\code{objective}}{}
}}

\keyword{datasets}
