% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectPostDTO.r
\name{ProjectPostDTO}
\alias{ProjectPostDTO}
\title{ProjectPostDTO Class}
\description{
ProjectPostDTO Class

ProjectPostDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{relatedProjects}}{}

\item{\code{financialFunding}}{}

\item{\code{financialReference}}{}

\item{\code{description}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{keywords}}{}

\item{\code{homePage}}{}

\item{\code{administrativeContacts}}{}

\item{\code{coordinators}}{}

\item{\code{scientificContacts}}{}

\item{\code{objective}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectPostDTO$new()}}
\item \href{#method-toJSON}{\code{ProjectPostDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectPostDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ProjectPostDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ProjectPostDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProjectPostDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ProjectPostDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPostDTO$new(
  name,
  shortname,
  relatedProjects,
  financialFunding,
  financialReference,
  description,
  startDate,
  endDate,
  keywords,
  homePage,
  administrativeContacts,
  coordinators,
  scientificContacts,
  objective
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPostDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPostDTO$fromJSON(ProjectPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPostDTO$fromJSONObject(ProjectPostDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPostDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPostDTO$fromJSONString(ProjectPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPostDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
