% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectPutDTO.r
\docType{data}
\name{ProjectPutDTO}
\alias{ProjectPutDTO}
\title{ProjectPutDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectPutDTO
}
\description{
ProjectPutDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{relatedProjects}}{}

\item{\code{financialFunding}}{}

\item{\code{financialReference}}{}

\item{\code{description}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{keywords}}{}

\item{\code{homePage}}{}

\item{\code{administrativeContacts}}{}

\item{\code{coordinators}}{}

\item{\code{scientificContacts}}{}

\item{\code{objective}}{}
}}

\keyword{datasets}
