% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectsApi.r
\name{ProjectsApi}
\alias{ProjectsApi}
\title{Projects operations}
\description{
phisWSClientRTools.Projects
}
\section{Methods}{

\describe{

get_by_id Get a project by its URI.


get_by_search Get all projects corresponding to the searched params given


post3 Post a project


put2 Put project(s)

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectsApi$new()}}
\item \href{#method-get_by_id}{\code{ProjectsApi$get_by_id()}}
\item \href{#method-get_by_search}{\code{ProjectsApi$get_by_search()}}
\item \href{#method-post3}{\code{ProjectsApi$post3()}}
\item \href{#method-put2}{\code{ProjectsApi$put2()}}
\item \href{#method-clone}{\code{ProjectsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_by_id"></a>}}
\subsection{Method \code{get_by_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$get_by_id(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_by_search"></a>}}
\subsection{Method \code{get_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$get_by_search(
  page_size,
  page,
  uri,
  name,
  shortname,
  financial_funding,
  financial_reference,
  description,
  start_date,
  end_date,
  home_page,
  objective,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post3"></a>}}
\subsection{Method \code{post3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$post3(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put2"></a>}}
\subsection{Method \code{put2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$put2(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
