% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyDTO.r
\docType{data}
\name{PropertyDTO}
\alias{PropertyDTO}
\title{PropertyDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PropertyDTO
}
\description{
PropertyDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{relation}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
