% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvenanceDTO.r
\docType{data}
\name{ProvenanceDTO}
\alias{ProvenanceDTO}
\title{ProvenanceDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProvenanceDTO
}
\description{
ProvenanceDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
