% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvenancePostDTO.r
\docType{data}
\name{ProvenancePostDTO}
\alias{ProvenancePostDTO}
\title{ProvenancePostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProvenancePostDTO
}
\description{
ProvenancePostDTO Class
}
\section{Fields}{

\describe{
\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
