% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvenancesApi.r
\name{ProvenancesApi}
\alias{ProvenancesApi}
\title{Provenances operations}
\description{
phisWSClientRTools.Provenances
}
\section{Methods}{

\describe{

get_provenances Get all provenances corresponding to the search params given


post4 Post provenance(s)


put3 Update provenance

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProvenancesApi$new()}}
\item \href{#method-get_provenances}{\code{ProvenancesApi$get_provenances()}}
\item \href{#method-post4}{\code{ProvenancesApi$post4()}}
\item \href{#method-put3}{\code{ProvenancesApi$put3()}}
\item \href{#method-clone}{\code{ProvenancesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvenancesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_provenances"></a>}}
\subsection{Method \code{get_provenances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvenancesApi$get_provenances(
  page_size,
  page,
  uri,
  label,
  comment,
  json_value_filter,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post4"></a>}}
\subsection{Method \code{post4()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvenancesApi$post4(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put3"></a>}}
\subsection{Method \code{put3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvenancesApi$put3(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProvenancesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
