% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiometricTargetDTO.r
\docType{data}
\name{RadiometricTargetDTO}
\alias{RadiometricTargetDTO}
\title{RadiometricTargetDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RadiometricTargetDTO
}
\description{
RadiometricTargetDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
