% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiometricTargetPostDTO.r
\name{RadiometricTargetPostDTO}
\alias{RadiometricTargetPostDTO}
\title{RadiometricTargetPostDTO Class}
\description{
RadiometricTargetPostDTO Class

RadiometricTargetPostDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{label}}{}

\item{\code{properties}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RadiometricTargetPostDTO$new()}}
\item \href{#method-toJSON}{\code{RadiometricTargetPostDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{RadiometricTargetPostDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{RadiometricTargetPostDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{RadiometricTargetPostDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RadiometricTargetPostDTO$fromJSONString()}}
\item \href{#method-clone}{\code{RadiometricTargetPostDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetPostDTO$new(label, properties)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetPostDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetPostDTO$fromJSON(RadiometricTargetPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetPostDTO$fromJSONObject(RadiometricTargetPostDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetPostDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetPostDTO$fromJSONString(RadiometricTargetPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetPostDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
