% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiometricTargetsApi.r
\name{RadiometricTargetsApi}
\alias{RadiometricTargetsApi}
\title{RadiometricTargets operations}
\description{
phisWSClientRTools.RadiometricTargets
}
\section{Methods}{

\describe{

get_radiometric_targets_by_search Get all radiometric targets corresponding to the search params given


get_radiometric_targets_details Get all radiometric target&#39;s details corresponding to the search uri


post_radiometric_targets Post radiometric(s) target(s) 


put4 Update radiometric targets

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RadiometricTargetsApi$new()}}
\item \href{#method-get_radiometric_targets_by_search}{\code{RadiometricTargetsApi$get_radiometric_targets_by_search()}}
\item \href{#method-get_radiometric_targets_details}{\code{RadiometricTargetsApi$get_radiometric_targets_details()}}
\item \href{#method-post_radiometric_targets}{\code{RadiometricTargetsApi$post_radiometric_targets()}}
\item \href{#method-put4}{\code{RadiometricTargetsApi$put4()}}
\item \href{#method-clone}{\code{RadiometricTargetsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_radiometric_targets_by_search"></a>}}
\subsection{Method \code{get_radiometric_targets_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetsApi$get_radiometric_targets_by_search(
  page_size,
  page,
  uri,
  label,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_radiometric_targets_details"></a>}}
\subsection{Method \code{get_radiometric_targets_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetsApi$get_radiometric_targets_details(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_radiometric_targets"></a>}}
\subsection{Method \code{post_radiometric_targets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetsApi$post_radiometric_targets(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put4"></a>}}
\subsection{Method \code{put4()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetsApi$put4(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
