% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdfResourceDefinitionDTO.r
\name{RdfResourceDefinitionDTO}
\alias{RdfResourceDefinitionDTO}
\title{RdfResourceDefinitionDTO Class}
\description{
RdfResourceDefinitionDTO Class

RdfResourceDefinitionDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{properties}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RdfResourceDefinitionDTO$new()}}
\item \href{#method-toJSON}{\code{RdfResourceDefinitionDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{RdfResourceDefinitionDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{RdfResourceDefinitionDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{RdfResourceDefinitionDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RdfResourceDefinitionDTO$fromJSONString()}}
\item \href{#method-clone}{\code{RdfResourceDefinitionDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfResourceDefinitionDTO$new(uri, label, properties)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfResourceDefinitionDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfResourceDefinitionDTO$fromJSON(RdfResourceDefinitionDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfResourceDefinitionDTO$fromJSONObject(RdfResourceDefinitionDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfResourceDefinitionDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfResourceDefinitionDTO$fromJSONString(RdfResourceDefinitionDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RdfResourceDefinitionDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
