% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseFormPOST.r
\docType{data}
\name{ResponseFormPOST}
\alias{ResponseFormPOST}
\title{ResponseFormPOST Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResponseFormPOST
}
\description{
ResponseFormPOST Class
}
\section{Fields}{

\describe{
\item{\code{metadata}}{}
}}

\keyword{datasets}
