% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseUnique.r
\docType{data}
\name{ResponseUnique}
\alias{ResponseUnique}
\title{ResponseUnique Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResponseUnique
}
\description{
ResponseUnique Class
}
\keyword{datasets}
