% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectDTO.r
\name{ScientificObjectDTO}
\alias{ScientificObjectDTO}
\title{ScientificObjectDTO Class}
\description{
ScientificObjectDTO Class

ScientificObjectDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{geometry}}{}

\item{\code{experiment}}{}

\item{\code{isPartOf}}{}

\item{\code{label}}{}

\item{\code{properties}}{}

\item{\code{uriExperiment}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectDTO$new()}}
\item \href{#method-toJSON}{\code{ScientificObjectDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ScientificObjectDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ScientificObjectDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ScientificObjectDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScientificObjectDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ScientificObjectDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDTO$new(
  uri,
  rdfType,
  geometry,
  experiment,
  isPartOf,
  label,
  properties,
  uriExperiment
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDTO$fromJSON(ScientificObjectDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDTO$fromJSONObject(ScientificObjectDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDTO$fromJSONString(ScientificObjectDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
