% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectPostDTO.r
\name{ScientificObjectPostDTO}
\alias{ScientificObjectPostDTO}
\title{ScientificObjectPostDTO Class}
\description{
ScientificObjectPostDTO Class

ScientificObjectPostDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rdfType}}{}

\item{\code{geometry}}{}

\item{\code{experiment}}{}

\item{\code{isPartOf}}{}

\item{\code{year}}{}

\item{\code{properties}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectPostDTO$new()}}
\item \href{#method-toJSON}{\code{ScientificObjectPostDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ScientificObjectPostDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ScientificObjectPostDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ScientificObjectPostDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScientificObjectPostDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ScientificObjectPostDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPostDTO$new(
  rdfType,
  geometry,
  experiment,
  isPartOf,
  year,
  properties
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPostDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPostDTO$fromJSON(ScientificObjectPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPostDTO$fromJSONObject(ScientificObjectPostDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPostDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPostDTO$fromJSONString(ScientificObjectPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPostDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
