% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectPutDTO.r
\name{ScientificObjectPutDTO}
\alias{ScientificObjectPutDTO}
\title{ScientificObjectPutDTO Class}
\description{
ScientificObjectPutDTO Class

ScientificObjectPutDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rdfType}}{}

\item{\code{geometry}}{}

\item{\code{isPartOf}}{}

\item{\code{label}}{}

\item{\code{properties}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectPutDTO$new()}}
\item \href{#method-toJSON}{\code{ScientificObjectPutDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ScientificObjectPutDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ScientificObjectPutDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ScientificObjectPutDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScientificObjectPutDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ScientificObjectPutDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPutDTO$new(rdfType, geometry, isPartOf, label, properties)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPutDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPutDTO$fromJSON(ScientificObjectPutDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPutDTO$fromJSONObject(ScientificObjectPutDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPutDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPutDTO$fromJSONString(ScientificObjectPutDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectPutDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
