% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectsApi.r
\docType{data}
\name{ScientificObjectsApi}
\alias{ScientificObjectsApi}
\title{ScientificObjects operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectsApi
}
\description{
phisWSClientRTools.ScientificObjects
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_scientific_objects_by_search Get all scientific objects corresponding to the searched params given


post_scientific_object Post scientific object(s)


put5 Put scientific object(s) in the given experiment

}
}

\keyword{datasets}
