% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorProfileDTO.r
\docType{data}
\name{SensorProfileDTO}
\alias{SensorProfileDTO}
\title{SensorProfileDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SensorProfileDTO
}
\description{
SensorProfileDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
