% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorsApi.r
\docType{data}
\name{SensorsApi}
\alias{SensorsApi}
\title{Sensors operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SensorsApi
}
\description{
opensilexWSClient.Sensors
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_sensor_data Get data mesured by a sensor


get_sensor_details Get a sensor


get_sensor_profile Get a sensor profile


get_sensors_by_search Get all sensors corresponding to the search params given


post6 Post a sensor


post_profiles Post sensor(s) profile(s)


put6 Update sensors


put_measured_variables1 Update the measured variables of a sensor

}
}

\keyword{datasets}
