% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShootingConfigurationDTO.r
\name{ShootingConfigurationDTO}
\alias{ShootingConfigurationDTO}
\title{ShootingConfigurationDTO Class}
\description{
ShootingConfigurationDTO Class

ShootingConfigurationDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{date}}{}

\item{\code{position}}{}

\item{\code{sensor}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShootingConfigurationDTO$new()}}
\item \href{#method-toJSON}{\code{ShootingConfigurationDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ShootingConfigurationDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ShootingConfigurationDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ShootingConfigurationDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ShootingConfigurationDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ShootingConfigurationDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShootingConfigurationDTO$new(date, position, sensor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShootingConfigurationDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShootingConfigurationDTO$fromJSON(ShootingConfigurationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShootingConfigurationDTO$fromJSONObject(ShootingConfigurationDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShootingConfigurationDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShootingConfigurationDTO$fromJSONString(ShootingConfigurationDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShootingConfigurationDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
