% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Status.r
\docType{data}
\name{Status}
\alias{Status}
\title{Status Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Status
}
\description{
Status Class
}
\section{Fields}{

\describe{
\item{\code{message}}{}

\item{\code{exception}}{}
}}

\keyword{datasets}
