% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyDetails.r
\docType{data}
\name{StudyDetails}
\alias{StudyDetails}
\title{StudyDetails Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StudyDetails
}
\description{
StudyDetails Class
}
\section{Fields}{

\describe{
\item{\code{studyDbId}}{}

\item{\code{studyName}}{}

\item{\code{studyTypeDbId}}{}

\item{\code{studyTypeName}}{}

\item{\code{studyDescription}}{}

\item{\code{seasons}}{}

\item{\code{commonCropName}}{}

\item{\code{trialDbId}}{}

\item{\code{trialName}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{active}}{}

\item{\code{license}}{}

\item{\code{location}}{}

\item{\code{contacts}}{}

\item{\code{dataLinks}}{}

\item{\code{lastUpdate}}{}

\item{\code{additionalInfo}}{}

\item{\code{documentationURL}}{}

\item{\code{studyType}}{}
}}

\keyword{datasets}
