% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySearchDTO.r
\name{StudySearchDTO}
\alias{StudySearchDTO}
\title{StudySearchDTO Class}
\description{
StudySearchDTO Class

StudySearchDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{commonCropName}}{}

\item{\code{seasonDbId}}{}

\item{\code{studyDbIds}}{}

\item{\code{studyNames}}{}

\item{\code{sortBy}}{}

\item{\code{sortOrder}}{}

\item{\code{page}}{}

\item{\code{pageSize}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StudySearchDTO$new()}}
\item \href{#method-toJSON}{\code{StudySearchDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{StudySearchDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{StudySearchDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{StudySearchDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StudySearchDTO$fromJSONString()}}
\item \href{#method-clone}{\code{StudySearchDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudySearchDTO$new(
  commonCropName,
  seasonDbId,
  studyDbIds,
  studyNames,
  sortBy,
  sortOrder,
  page,
  pageSize
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudySearchDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudySearchDTO$fromJSON(StudySearchDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudySearchDTO$fromJSONObject(StudySearchDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudySearchDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudySearchDTO$fromJSONString(StudySearchDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudySearchDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
