% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trait.r
\docType{data}
\name{Trait}
\alias{Trait}
\title{Trait Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Trait
}
\description{
Trait Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
