% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitDTO.r
\docType{data}
\name{TraitDTO}
\alias{TraitDTO}
\title{TraitDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TraitDTO
}
\description{
TraitDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}
}}

\keyword{datasets}
