% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitsApi.r
\name{TraitsApi}
\alias{TraitsApi}
\title{Traits operations}
\description{
phisWSClientRTools.Traits
}
\section{Methods}{

\describe{

get_trait_details Get a trait


get_traits_by_search Get all Traits corresponding to the searched params given


post_trait Post trait(s)


put_trait Update trait

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TraitsApi$new()}}
\item \href{#method-get_trait_details}{\code{TraitsApi$get_trait_details()}}
\item \href{#method-get_traits_by_search}{\code{TraitsApi$get_traits_by_search()}}
\item \href{#method-post_trait}{\code{TraitsApi$post_trait()}}
\item \href{#method-put_trait}{\code{TraitsApi$put_trait()}}
\item \href{#method-clone}{\code{TraitsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_trait_details"></a>}}
\subsection{Method \code{get_trait_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitsApi$get_trait_details(trait, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_traits_by_search"></a>}}
\subsection{Method \code{get_traits_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitsApi$get_traits_by_search(page_size, page, uri, label, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_trait"></a>}}
\subsection{Method \code{post_trait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitsApi$post_trait(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_trait"></a>}}
\subsection{Method \code{put_trait()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitsApi$put_trait(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TraitsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
