% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unit.r
\docType{data}
\name{Unit}
\alias{Unit}
\title{Unit Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Unit
}
\description{
Unit Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
