% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitDTO.r
\docType{data}
\name{UnitDTO}
\alias{UnitDTO}
\title{UnitDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnitDTO
}
\description{
UnitDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}
}}

\keyword{datasets}
