% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsApi.r
\name{UnitsApi}
\alias{UnitsApi}
\title{Units operations}
\description{
phisWSClientRTools.Units
}
\section{Methods}{

\describe{

get_unit_details Get a unit


get_units_by_search Get all units corresponding to the searched params given


post_unit Post unit(s)


put_unit Update unit

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UnitsApi$new()}}
\item \href{#method-get_unit_details}{\code{UnitsApi$get_unit_details()}}
\item \href{#method-get_units_by_search}{\code{UnitsApi$get_units_by_search()}}
\item \href{#method-post_unit}{\code{UnitsApi$post_unit()}}
\item \href{#method-put_unit}{\code{UnitsApi$put_unit()}}
\item \href{#method-clone}{\code{UnitsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_unit_details"></a>}}
\subsection{Method \code{get_unit_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitsApi$get_unit_details(unit, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_units_by_search"></a>}}
\subsection{Method \code{get_units_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitsApi$get_units_by_search(page_size, page, uri, label, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_unit"></a>}}
\subsection{Method \code{post_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitsApi$post_unit(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_unit"></a>}}
\subsection{Method \code{put_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitsApi$put_unit(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
