% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uri.r
\docType{data}
\name{Uri}
\alias{Uri}
\title{Uri Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Uri
}
\description{
Uri Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{properties}}{}

\item{\code{annotations}}{}
}}

\keyword{datasets}
