% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UriApi.r
\name{UriApi}
\alias{UriApi}
\title{Uri operations}
\description{
phisWSClientRTools.Uri
}
\section{Methods}{

\describe{

get_ancestors Get all the ancestor of an uri


get_descendants Get all the descendants of an uri


get_instances_by_concept Get all the instances of a concept


get_sibblings Get all the siblings of an Uri


get_type_if_uri_exist get the type of an uri if it exist


get_uri_metadata Get all uri informations


get_uris_by_label get all uri with a given label


is_uri_existing check if an uri exists or not (in the triplestore)

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UriApi$new()}}
\item \href{#method-get_ancestors}{\code{UriApi$get_ancestors()}}
\item \href{#method-get_descendants}{\code{UriApi$get_descendants()}}
\item \href{#method-get_instances_by_concept}{\code{UriApi$get_instances_by_concept()}}
\item \href{#method-get_sibblings}{\code{UriApi$get_sibblings()}}
\item \href{#method-get_type_if_uri_exist}{\code{UriApi$get_type_if_uri_exist()}}
\item \href{#method-get_uri_metadata}{\code{UriApi$get_uri_metadata()}}
\item \href{#method-get_uris_by_label}{\code{UriApi$get_uris_by_label()}}
\item \href{#method-is_uri_existing}{\code{UriApi$is_uri_existing()}}
\item \href{#method-clone}{\code{UriApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ancestors"></a>}}
\subsection{Method \code{get_ancestors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$get_ancestors(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_descendants"></a>}}
\subsection{Method \code{get_descendants()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$get_descendants(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_instances_by_concept"></a>}}
\subsection{Method \code{get_instances_by_concept()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$get_instances_by_concept(uri, deep, language, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sibblings"></a>}}
\subsection{Method \code{get_sibblings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$get_sibblings(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_type_if_uri_exist"></a>}}
\subsection{Method \code{get_type_if_uri_exist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$get_type_if_uri_exist(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uri_metadata"></a>}}
\subsection{Method \code{get_uri_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$get_uri_metadata(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uris_by_label"></a>}}
\subsection{Method \code{get_uris_by_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$get_uris_by_label(label, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_uri_existing"></a>}}
\subsection{Method \code{is_uri_existing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$is_uri_existing(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UriApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
