% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.r
\docType{data}
\name{User}
\alias{User}
\title{User Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
User
}
\description{
User Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{familyName}}{}

\item{\code{phone}}{}

\item{\code{affiliation}}{}

\item{\code{uri}}{}

\item{\code{orcid}}{}

\item{\code{groups}}{}
}}

\keyword{datasets}
