% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserApi.r
\name{UserApi}
\alias{UserApi}
\title{User operations}
\description{
phisWSClientRTools.User
}
\section{Methods}{

\describe{

get_user_by_search Get all users corresponding to the searched params given


get_user_details Get a user


post_user Post a user


put_user Update users

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserApi$new()}}
\item \href{#method-get_user_by_search}{\code{UserApi$get_user_by_search()}}
\item \href{#method-get_user_details}{\code{UserApi$get_user_details()}}
\item \href{#method-post_user}{\code{UserApi$post_user()}}
\item \href{#method-put_user}{\code{UserApi$put_user()}}
\item \href{#method-clone}{\code{UserApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_user_by_search"></a>}}
\subsection{Method \code{get_user_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$get_user_by_search(
  page_size,
  page,
  email,
  first_name,
  family_name,
  address,
  phone,
  affiliation,
  orcid,
  admin,
  available,
  uri,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_user_details"></a>}}
\subsection{Method \code{get_user_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$get_user_details(user_email, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_user"></a>}}
\subsection{Method \code{post_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$post_user(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_user"></a>}}
\subsection{Method \code{put_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$put_user(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
