% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserDTO.r
\name{UserDTO}
\alias{UserDTO}
\title{UserDTO Class}
\description{
UserDTO Class

UserDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{}

\item{\code{password}}{}

\item{\code{firstName}}{}

\item{\code{familyName}}{}

\item{\code{address}}{}

\item{\code{phone}}{}

\item{\code{affiliation}}{}

\item{\code{orcid}}{}

\item{\code{admin}}{}

\item{\code{groupsUris}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UserDTO$new()}}
\item \href{#method-toJSON}{\code{UserDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{UserDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{UserDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{UserDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{UserDTO$fromJSONString()}}
\item \href{#method-clone}{\code{UserDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDTO$new(
  email,
  password,
  firstName,
  familyName,
  address,
  phone,
  affiliation,
  orcid,
  admin,
  groupsUris
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDTO$fromJSON(UserDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDTO$fromJSONObject(UserDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDTO$fromJSONString(UserDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
