% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableDTO.r
\docType{data}
\name{VariableDTO}
\alias{VariableDTO}
\title{VariableDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariableDTO
}
\description{
VariableDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}

\item{\code{trait}}{}

\item{\code{method}}{}

\item{\code{unit}}{}
}}

\keyword{datasets}
