% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesApi.r
\name{VariablesApi}
\alias{VariablesApi}
\title{Variables operations}
\description{
phisWSClientRTools.Variables
}
\section{Methods}{

\describe{

get_variable_detail Get a variable


get_variables_by_search Get all variables corresponding to the searched params given


get_variables_details_by_search Get all variables details corresponding to the searched params given


post_variable Post variable(s)


put_variable Update variable

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariablesApi$new()}}
\item \href{#method-get_variable_detail}{\code{VariablesApi$get_variable_detail()}}
\item \href{#method-get_variables_by_search}{\code{VariablesApi$get_variables_by_search()}}
\item \href{#method-get_variables_details_by_search}{\code{VariablesApi$get_variables_details_by_search()}}
\item \href{#method-post_variable}{\code{VariablesApi$post_variable()}}
\item \href{#method-put_variable}{\code{VariablesApi$put_variable()}}
\item \href{#method-clone}{\code{VariablesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variable_detail"></a>}}
\subsection{Method \code{get_variable_detail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_variable_detail(variable, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variables_by_search"></a>}}
\subsection{Method \code{get_variables_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_variables_by_search(
  page_size,
  page,
  uri,
  label,
  trait,
  trait_s_kos_reference,
  method,
  unit,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variables_details_by_search"></a>}}
\subsection{Method \code{get_variables_details_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$get_variables_details_by_search(
  page_size,
  page,
  uri,
  label,
  trait,
  trait_s_kos_reference,
  method,
  unit,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_variable"></a>}}
\subsection{Method \code{post_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$post_variable(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_variable"></a>}}
\subsection{Method \code{put_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$put_variable(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariablesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
