% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vector.r
\docType{data}
\name{Vector}
\alias{Vector}
\title{Vector Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Vector
}
\description{
Vector Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{personInCharge}}{}
}}

\keyword{datasets}
