% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDTO.r
\name{VectorDTO}
\alias{VectorDTO}
\title{VectorDTO Class}
\description{
VectorDTO Class

VectorDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{personInCharge}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VectorDTO$new()}}
\item \href{#method-toJSON}{\code{VectorDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{VectorDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{VectorDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{VectorDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VectorDTO$fromJSONString()}}
\item \href{#method-clone}{\code{VectorDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDTO$new(
  uri,
  rdfType,
  label,
  brand,
  serialNumber,
  inServiceDate,
  dateOfPurchase,
  personInCharge
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDTO$fromJSON(VectorDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDTO$fromJSONObject(VectorDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDTO$fromJSONString(VectorDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
