% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorsApi.r
\name{VectorsApi}
\alias{VectorsApi}
\title{Vectors operations}
\description{
phisWSClientRTools.Vectors
}
\section{Methods}{

\describe{

get_vector_details Get a vector


get_vectors_by_search Get all vectors corresponding to the search params given


post6 Post a vector


put7 Update vector

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VectorsApi$new()}}
\item \href{#method-get_vector_details}{\code{VectorsApi$get_vector_details()}}
\item \href{#method-get_vectors_by_search}{\code{VectorsApi$get_vectors_by_search()}}
\item \href{#method-post6}{\code{VectorsApi$post6()}}
\item \href{#method-put7}{\code{VectorsApi$put7()}}
\item \href{#method-clone}{\code{VectorsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vector_details"></a>}}
\subsection{Method \code{get_vector_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorsApi$get_vector_details(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vectors_by_search"></a>}}
\subsection{Method \code{get_vectors_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorsApi$get_vectors_by_search(
  page_size,
  page,
  uri,
  rdf_type,
  label,
  brand,
  serial_number,
  in_service_date,
  date_of_purchase,
  person_in_charge,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post6"></a>}}
\subsection{Method \code{post6()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorsApi$post6(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put7"></a>}}
\subsection{Method \code{put7()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorsApi$put7(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
