% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabulariesApi.r
\name{VocabulariesApi}
\alias{VocabulariesApi}
\title{Vocabularies operations}
\description{
phisWSClientRTools.Vocabularies
}
\section{Methods}{

\describe{

get_contacts Get all contact properties that can be added to a given rdfType


get_device_properties Get all device properties that can be added to a given rdfType


get_namespaces Get all triplestore namespaces


get_rdfs Get all rdfs properties that can be added to an instance

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VocabulariesApi$new()}}
\item \href{#method-get_contacts}{\code{VocabulariesApi$get_contacts()}}
\item \href{#method-get_device_properties}{\code{VocabulariesApi$get_device_properties()}}
\item \href{#method-get_namespaces}{\code{VocabulariesApi$get_namespaces()}}
\item \href{#method-get_rdfs}{\code{VocabulariesApi$get_rdfs()}}
\item \href{#method-clone}{\code{VocabulariesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VocabulariesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_contacts"></a>}}
\subsection{Method \code{get_contacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VocabulariesApi$get_contacts(rdf_type, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_device_properties"></a>}}
\subsection{Method \code{get_device_properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VocabulariesApi$get_device_properties(rdf_type, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_namespaces"></a>}}
\subsection{Method \code{get_namespaces()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VocabulariesApi$get_namespaces(page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rdfs"></a>}}
\subsection{Method \code{get_rdfs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VocabulariesApi$get_rdfs(page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VocabulariesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
