% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WasGeneratedBy.r
\docType{data}
\name{WasGeneratedBy}
\alias{WasGeneratedBy}
\title{WasGeneratedBy Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WasGeneratedBy
}
\description{
WasGeneratedBy Class
}
\section{Fields}{

\describe{
\item{\code{wasGeneratedBy}}{}

\item{\code{wasGeneratedByDescription}}{}
}}

\keyword{datasets}
