% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WasGeneratedByDTO.r
\docType{data}
\name{WasGeneratedByDTO}
\alias{WasGeneratedByDTO}
\title{WasGeneratedByDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WasGeneratedByDTO
}
\description{
WasGeneratedByDTO Class
}
\section{Fields}{

\describe{
\item{\code{wasGeneratedByDocument}}{}

\item{\code{wasGeneratedByDescription}}{}
}}

\keyword{datasets}
