% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcquisitionSessionsApi.r
\docType{data}
\name{AcquisitionSessionsApi}
\alias{AcquisitionSessionsApi}
\title{AcquisitionSessions operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AcquisitionSessionsApi
}
\description{
phisWSClientRTools.AcquisitionSessions
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get4_p_metadata_file Get the metadata for the acquisition session file

}
}

\keyword{datasets}
