% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActuatorDetailDTO.r
\name{ActuatorDetailDTO}
\alias{ActuatorDetailDTO}
\title{ActuatorDetailDTO Class}
\description{
ActuatorDetailDTO Class

ActuatorDetailDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{model}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{dateOfLastCalibration}}{}

\item{\code{personInCharge}}{}

\item{\code{variables}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ActuatorDetailDTO$new()}}
\item \href{#method-toJSON}{\code{ActuatorDetailDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ActuatorDetailDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ActuatorDetailDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ActuatorDetailDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ActuatorDetailDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ActuatorDetailDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorDetailDTO$new(
  uri,
  rdfType,
  label,
  brand,
  model,
  serialNumber,
  inServiceDate,
  dateOfPurchase,
  dateOfLastCalibration,
  personInCharge,
  variables
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorDetailDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorDetailDTO$fromJSON(ActuatorDetailDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorDetailDTO$fromJSONObject(ActuatorDetailDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorDetailDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorDetailDTO$fromJSONString(ActuatorDetailDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActuatorDetailDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
