% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActuatorDTO.r
\docType{data}
\name{ActuatorDTO}
\alias{ActuatorDTO}
\title{ActuatorDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ActuatorDTO
}
\description{
ActuatorDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{model}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{dateOfLastCalibration}}{}

\item{\code{personInCharge}}{}
}}

\keyword{datasets}
