% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActuatorsApi.r
\docType{data}
\name{ActuatorsApi}
\alias{ActuatorsApi}
\title{Actuators operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ActuatorsApi
}
\description{
opensilexWSClient.Actuators
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_actuator_details Get an actuator


get_actuators_by_search Get all actuators corresponding to the search params given


post Post actuator(s)


put Put actuator(s)


put_measured_variables Update the measured variables of an actuator

}
}

\keyword{datasets}
