% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditionalInfo.r
\docType{data}
\name{AdditionalInfo}
\alias{AdditionalInfo}
\title{AdditionalInfo Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AdditionalInfo
}
\description{
AdditionalInfo Class
}
\section{Fields}{

\describe{
\item{\code{keywords}}{}

\item{\code{projectsNames}}{}
}}

\keyword{datasets}
