% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgronomicalData.r
\name{AgronomicalData}
\alias{AgronomicalData}
\title{AgronomicalData Class}
\description{
AgronomicalData Class

AgronomicalData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{agronomicalObject}}{}

\item{\code{date}}{}

\item{\code{value}}{}

\item{\code{variable}}{}

\item{\code{sensor}}{}

\item{\code{incertitude}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AgronomicalData$new()}}
\item \href{#method-toJSON}{\code{AgronomicalData$toJSON()}}
\item \href{#method-fromJSON}{\code{AgronomicalData$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{AgronomicalData$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{AgronomicalData$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AgronomicalData$fromJSONString()}}
\item \href{#method-clone}{\code{AgronomicalData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalData$new(
  agronomicalObject,
  date,
  value,
  variable,
  sensor,
  incertitude
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalData$fromJSON(AgronomicalDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalData$fromJSONObject(AgronomicalDataObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalData$fromJSONString(AgronomicalDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
