% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgronomicalDataDTO.r
\name{AgronomicalDataDTO}
\alias{AgronomicalDataDTO}
\title{AgronomicalDataDTO Class}
\description{
AgronomicalDataDTO Class

AgronomicalDataDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{agronomicalObject}}{}

\item{\code{date}}{}

\item{\code{value}}{}

\item{\code{sensor}}{}

\item{\code{incertitude}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AgronomicalDataDTO$new()}}
\item \href{#method-toJSON}{\code{AgronomicalDataDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{AgronomicalDataDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{AgronomicalDataDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{AgronomicalDataDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AgronomicalDataDTO$fromJSONString()}}
\item \href{#method-clone}{\code{AgronomicalDataDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalDataDTO$new(agronomicalObject, date, value, sensor, incertitude)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalDataDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalDataDTO$fromJSON(AgronomicalDataDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalDataDTO$fromJSONObject(AgronomicalDataDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalDataDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalDataDTO$fromJSONString(AgronomicalDataDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgronomicalDataDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
