% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPostDTO.r
\name{AnnotationPostDTO}
\alias{AnnotationPostDTO}
\title{AnnotationPostDTO Class}
\description{
AnnotationPostDTO Class

AnnotationPostDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{creator}}{}

\item{\code{motivatedBy}}{}

\item{\code{bodyValues}}{}

\item{\code{targets}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnnotationPostDTO$new()}}
\item \href{#method-toJSON}{\code{AnnotationPostDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{AnnotationPostDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{AnnotationPostDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{AnnotationPostDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AnnotationPostDTO$fromJSONString()}}
\item \href{#method-clone}{\code{AnnotationPostDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPostDTO$new(creator, motivatedBy, bodyValues, targets)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPostDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPostDTO$fromJSON(AnnotationPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPostDTO$fromJSONObject(AnnotationPostDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPostDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPostDTO$fromJSONString(AnnotationPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPostDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
