% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationsApi.r
\name{AnnotationsApi}
\alias{AnnotationsApi}
\title{Annotations operations}
\description{
phisWSClientRTools.Annotations
}
\section{Methods}{

\describe{

get_annotation_by_uri Get a annotation


get_annotations_by_search Get all annotations corresponding to the search params given


post1 Post annotations

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnnotationsApi$new()}}
\item \href{#method-get_annotation_by_uri}{\code{AnnotationsApi$get_annotation_by_uri()}}
\item \href{#method-get_annotations_by_search}{\code{AnnotationsApi$get_annotations_by_search()}}
\item \href{#method-post1}{\code{AnnotationsApi$post1()}}
\item \href{#method-clone}{\code{AnnotationsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_annotation_by_uri"></a>}}
\subsection{Method \code{get_annotation_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$get_annotation_by_uri(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_annotations_by_search"></a>}}
\subsection{Method \code{get_annotations_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$get_annotations_by_search(
  page_size,
  page,
  uri,
  creator,
  target,
  description,
  motivated_by,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post1"></a>}}
\subsection{Method \code{post1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$post1(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
