% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiDescriptionDTO.r
\docType{data}
\name{ApiDescriptionDTO}
\alias{ApiDescriptionDTO}
\title{ApiDescriptionDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiDescriptionDTO
}
\description{
ApiDescriptionDTO Class
}
\section{Fields}{

\describe{
\item{\code{buildVersion}}{}

\item{\code{majorVersion}}{}

\item{\code{name}}{}
}}

\keyword{datasets}
