% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ask.r
\docType{data}
\name{Ask}
\alias{Ask}
\title{Ask Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Ask
}
\description{
Ask Class
}
\section{Fields}{

\describe{
\item{\code{exist}}{}
}}

\keyword{datasets}
