% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiMethod.r
\docType{data}
\name{BrapiMethod}
\alias{BrapiMethod}
\title{BrapiMethod Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiMethod
}
\description{
BrapiMethod Class
}
\section{Fields}{

\describe{
\item{\code{methodDbId}}{}

\item{\code{methodName}}{}

\item{\code{brapiClass}}{}

\item{\code{description}}{}

\item{\code{formula}}{}

\item{\code{ontologyReference}}{}

\item{\code{reference}}{}
}}

\keyword{datasets}
