% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiObservationDTO.r
\docType{data}
\name{BrapiObservationDTO}
\alias{BrapiObservationDTO}
\title{BrapiObservationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiObservationDTO
}
\description{
BrapiObservationDTO Class
}
\section{Fields}{

\describe{
\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{observationDbId}}{}

\item{\code{observationLevel}}{}

\item{\code{observationTimeStamp}}{}

\item{\code{observationUnitDbId}}{}

\item{\code{observationUnitName}}{}

\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{operator}}{}

\item{\code{season}}{}

\item{\code{studyDbId}}{}

\item{\code{uploadedBy}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
