% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiObservationSummaryDTO.r
\name{BrapiObservationSummaryDTO}
\alias{BrapiObservationSummaryDTO}
\title{BrapiObservationSummaryDTO Class}
\description{
BrapiObservationSummaryDTO Class

BrapiObservationSummaryDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{collector}}{}

\item{\code{observationDbId}}{}

\item{\code{observationTimeStamp}}{}

\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{season}}{}

\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiObservationSummaryDTO$new()}}
\item \href{#method-toJSON}{\code{BrapiObservationSummaryDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiObservationSummaryDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiObservationSummaryDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiObservationSummaryDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiObservationSummaryDTO$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiObservationSummaryDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationSummaryDTO$new(
  collector,
  observationDbId,
  observationTimeStamp,
  observationVariableDbId,
  observationVariableName,
  season,
  value
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationSummaryDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationSummaryDTO$fromJSON(BrapiObservationSummaryDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationSummaryDTO$fromJSONObject(BrapiObservationSummaryDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationSummaryDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationSummaryDTO$fromJSONString(BrapiObservationSummaryDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationSummaryDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
