% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiObservationUnitDTO.r
\name{BrapiObservationUnitDTO}
\alias{BrapiObservationUnitDTO}
\title{BrapiObservationUnitDTO Class}
\description{
BrapiObservationUnitDTO Class

BrapiObservationUnitDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{blockNumber}}{}

\item{\code{entryNumber}}{}

\item{\code{entryType}}{}

\item{\code{germplasmDbId}}{}

\item{\code{germplasmName}}{}

\item{\code{locationDbId}}{}

\item{\code{locationName}}{}

\item{\code{observationLevel}}{}

\item{\code{observationLevels}}{}

\item{\code{observationUnitDbId}}{}

\item{\code{observationUnitName}}{}

\item{\code{observationUnitXref}}{}

\item{\code{observations}}{}

\item{\code{pedigree}}{}

\item{\code{plantNumber}}{}

\item{\code{plotNumber}}{}

\item{\code{positionCoordinateX}}{}

\item{\code{positionCoordinateXType}}{}

\item{\code{positionCoordinateY}}{}

\item{\code{positionCoordinateYType}}{}

\item{\code{programDbId}}{}

\item{\code{programName}}{}

\item{\code{replicate}}{}

\item{\code{studyDbId}}{}

\item{\code{studyName}}{}

\item{\code{treatments}}{}

\item{\code{trialDbId}}{}

\item{\code{trialName}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiObservationUnitDTO$new()}}
\item \href{#method-toJSON}{\code{BrapiObservationUnitDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiObservationUnitDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiObservationUnitDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiObservationUnitDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiObservationUnitDTO$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiObservationUnitDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationUnitDTO$new(
  blockNumber,
  entryNumber,
  entryType,
  germplasmDbId,
  germplasmName,
  locationDbId,
  locationName,
  observationLevel,
  observationLevels,
  observationUnitDbId,
  observationUnitName,
  observationUnitXref,
  observations,
  pedigree,
  plantNumber,
  plotNumber,
  positionCoordinateX,
  positionCoordinateXType,
  positionCoordinateY,
  positionCoordinateYType,
  programDbId,
  programName,
  replicate,
  studyDbId,
  studyName,
  treatments,
  trialDbId,
  trialName
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationUnitDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationUnitDTO$fromJSON(BrapiObservationUnitDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationUnitDTO$fromJSONObject(BrapiObservationUnitDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationUnitDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationUnitDTO$fromJSONString(BrapiObservationUnitDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiObservationUnitDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
