% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiScale.r
\docType{data}
\name{BrapiScale}
\alias{BrapiScale}
\title{BrapiScale Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiScale
}
\description{
BrapiScale Class
}
\section{Fields}{

\describe{
\item{\code{scaleDbid}}{}

\item{\code{scaleName}}{}

\item{\code{dataType}}{}

\item{\code{decimalPlaces}}{}

\item{\code{ontologyReference}}{}

\item{\code{xref}}{}

\item{\code{validValues}}{}
}}

\keyword{datasets}
