% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiTraitDTO.r
\name{BrapiTraitDTO}
\alias{BrapiTraitDTO}
\title{BrapiTraitDTO Class}
\description{
BrapiTraitDTO Class

BrapiTraitDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultValue}}{}

\item{\code{description}}{}

\item{\code{traitName}}{}

\item{\code{observationVariables}}{}

\item{\code{traitId}}{}

\item{\code{traitDbId}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiTraitDTO$new()}}
\item \href{#method-toJSON}{\code{BrapiTraitDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiTraitDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiTraitDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiTraitDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiTraitDTO$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiTraitDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiTraitDTO$new(
  defaultValue,
  description,
  traitName,
  observationVariables,
  traitId,
  traitDbId
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiTraitDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiTraitDTO$fromJSON(BrapiTraitDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiTraitDTO$fromJSONObject(BrapiTraitDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiTraitDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiTraitDTO$fromJSONString(BrapiTraitDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiTraitDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
