% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1callsApi.r
\docType{data}
\name{Brapiv1callsApi}
\alias{Brapiv1callsApi}
\title{Brapiv1calls operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Brapiv1callsApi
}
\description{
opensilexWSClient.Brapiv1calls
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_calls Check the available BrAPI calls

}
}

\keyword{datasets}
