% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1studiesApi.r
\docType{data}
\name{Brapiv1studiesApi}
\alias{Brapiv1studiesApi}
\title{Brapiv1studies operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Brapiv1studiesApi
}
\description{
phisWSClientRTools.Brapiv1studies
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_observation_units List all the observation units measured in the study.


get_observation_variables List all the observation variables measured in the study.


get_observations Get the observations associated to a specific study


get_studies Retrieve studies information


get_study_details Retrieve study details

}
}

\keyword{datasets}
