% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1studiesSearchApi.r
\name{Brapiv1studiesSearchApi}
\alias{Brapiv1studiesSearchApi}
\title{Brapiv1studiesSearch operations}
\description{
phisWSClientRTools.Brapiv1studiesSearch
}
\section{Methods}{

\describe{

get_studies_search Retrieve studies information


post_studies_search search studies

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Brapiv1studiesSearchApi$new()}}
\item \href{#method-get_studies_search}{\code{Brapiv1studiesSearchApi$get_studies_search()}}
\item \href{#method-post_studies_search}{\code{Brapiv1studiesSearchApi$post_studies_search()}}
\item \href{#method-clone}{\code{Brapiv1studiesSearchApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesSearchApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_studies_search"></a>}}
\subsection{Method \code{get_studies_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesSearchApi$get_studies_search(
  study_db_id,
  common_crop_name,
  season_db_id,
  active,
  sort_by,
  sort_order,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_studies_search"></a>}}
\subsection{Method \code{post_studies_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesSearchApi$post_studies_search(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Brapiv1studiesSearchApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
