% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1tokenApi.r
\docType{data}
\name{Brapiv1tokenApi}
\alias{Brapiv1tokenApi}
\title{Brapiv1token operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Brapiv1tokenApi
}
\description{
opensilexWSClient.Brapiv1token
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_token Login


log_out Log out

}
}

\keyword{datasets}
