% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1traitsApi.r
\docType{data}
\name{Brapiv1traitsApi}
\alias{Brapiv1traitsApi}
\title{Brapiv1traits operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Brapiv1traitsApi
}
\description{
phisWSClientRTools.Brapiv1traits
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_trait_details1 Retrieve trait details by id


get_traits_list Retrieve the list of all traits available in the system

}
}

\keyword{datasets}
