% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brapiv1variablesApi.r
\docType{data}
\name{Brapiv1variablesApi}
\alias{Brapiv1variablesApi}
\title{Brapiv1variables operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Brapiv1variablesApi
}
\description{
phisWSClientRTools.Brapiv1variables
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_observation_variables1 List all the observation variables measured in the study.


get_variable_details Retrieve variable details by id


get_variables_list Retrieve the list of all variables available in the system

}
}

\keyword{datasets}
