% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiVariable.r
\name{BrapiVariable}
\alias{BrapiVariable}
\title{BrapiVariable Class}
\description{
BrapiVariable Class

BrapiVariable Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{observationVariableDbId}}{}

\item{\code{observationVariableName}}{}

\item{\code{ontologyReference}}{}

\item{\code{synonyms}}{}

\item{\code{contextOfUse}}{}

\item{\code{growthStage}}{}

\item{\code{status}}{}

\item{\code{xref}}{}

\item{\code{institution}}{}

\item{\code{scientist}}{}

\item{\code{submissionTimesTamp}}{}

\item{\code{language}}{}

\item{\code{crop}}{}

\item{\code{trait}}{}

\item{\code{method}}{}

\item{\code{scale}}{}

\item{\code{defaultValue}}{}

\item{\code{documentationURL}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BrapiVariable$new()}}
\item \href{#method-toJSON}{\code{BrapiVariable$toJSON()}}
\item \href{#method-fromJSON}{\code{BrapiVariable$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{BrapiVariable$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{BrapiVariable$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BrapiVariable$fromJSONString()}}
\item \href{#method-clone}{\code{BrapiVariable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariable$new(
  observationVariableDbId,
  observationVariableName,
  ontologyReference,
  synonyms,
  contextOfUse,
  growthStage,
  status,
  xref,
  institution,
  scientist,
  submissionTimesTamp,
  language,
  crop,
  trait,
  method,
  scale,
  defaultValue,
  documentationURL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariable$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariable$fromJSON(BrapiVariableJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariable$fromJSONObject(BrapiVariableObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariable$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariable$fromJSONString(BrapiVariableJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrapiVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
