% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrapiVariableTrait.r
\docType{data}
\name{BrapiVariableTrait}
\alias{BrapiVariableTrait}
\title{BrapiVariableTrait Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BrapiVariableTrait
}
\description{
BrapiVariableTrait Class
}
\section{Fields}{

\describe{
\item{\code{traitDbId}}{}

\item{\code{traitName}}{}

\item{\code{brapiclass}}{}

\item{\code{description}}{}

\item{\code{synonyms}}{}

\item{\code{mainAbbreviation}}{}

\item{\code{alternativeAbbreviations}}{}

\item{\code{entity}}{}

\item{\code{attribute}}{}

\item{\code{status}}{}

\item{\code{xref}}{}

\item{\code{ontologyReference}}{}
}}

\keyword{datasets}
