% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcernedItem.r
\docType{data}
\name{ConcernedItem}
\alias{ConcernedItem}
\title{ConcernedItem Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConcernedItem
}
\description{
ConcernedItem Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{labels}}{}

\item{\code{objectLinked}}{}
}}

\keyword{datasets}
