% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcernedItemWithLabelsDTO.r
\name{ConcernedItemWithLabelsDTO}
\alias{ConcernedItemWithLabelsDTO}
\title{ConcernedItemWithLabelsDTO Class}
\description{
ConcernedItemWithLabelsDTO Class

ConcernedItemWithLabelsDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{typeURI}}{}

\item{\code{labels}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConcernedItemWithLabelsDTO$new()}}
\item \href{#method-toJSON}{\code{ConcernedItemWithLabelsDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ConcernedItemWithLabelsDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ConcernedItemWithLabelsDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ConcernedItemWithLabelsDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ConcernedItemWithLabelsDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ConcernedItemWithLabelsDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcernedItemWithLabelsDTO$new(uri, typeURI, labels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcernedItemWithLabelsDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcernedItemWithLabelsDTO$fromJSON(ConcernedItemWithLabelsDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcernedItemWithLabelsDTO$fromJSONObject(ConcernedItemWithLabelsDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcernedItemWithLabelsDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcernedItemWithLabelsDTO$fromJSONString(ConcernedItemWithLabelsDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConcernedItemWithLabelsDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
