% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactBrapi.r
\docType{data}
\name{ContactBrapi}
\alias{ContactBrapi}
\title{ContactBrapi Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContactBrapi
}
\description{
ContactBrapi Class
}
\section{Fields}{

\describe{
\item{\code{contactDbId}}{}

\item{\code{name}}{}

\item{\code{instituteName}}{}

\item{\code{email}}{}

\item{\code{type}}{}

\item{\code{orcid}}{}
}}

\keyword{datasets}
