% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactDTO.r
\docType{data}
\name{ContactDTO}
\alias{ContactDTO}
\title{ContactDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ContactDTO
}
\description{
ContactDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{firstname}}{}

\item{\code{lastname}}{}

\item{\code{email}}{}
}}

\keyword{datasets}
