% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactPostgreSQL.r
\name{ContactPostgreSQL}
\alias{ContactPostgreSQL}
\title{ContactPostgreSQL Class}
\description{
ContactPostgreSQL Class

ContactPostgreSQL Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{}

\item{\code{firstName}}{}

\item{\code{familyName}}{}

\item{\code{type}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ContactPostgreSQL$new()}}
\item \href{#method-toJSON}{\code{ContactPostgreSQL$toJSON()}}
\item \href{#method-fromJSON}{\code{ContactPostgreSQL$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ContactPostgreSQL$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ContactPostgreSQL$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ContactPostgreSQL$fromJSONString()}}
\item \href{#method-clone}{\code{ContactPostgreSQL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactPostgreSQL$new(email, firstName, familyName, type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactPostgreSQL$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactPostgreSQL$fromJSON(ContactPostgreSQLJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactPostgreSQL$fromJSONObject(ContactPostgreSQLObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactPostgreSQL$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactPostgreSQL$fromJSONString(ContactPostgreSQLJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContactPostgreSQL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
