% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.r
\docType{data}
\name{Data}
\alias{Data}
\title{Data Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data
}
\description{
Data Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{provenanceUri}}{}

\item{\code{objectUri}}{}

\item{\code{variableUri}}{}

\item{\code{date}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
