% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataApi.r
\docType{data}
\name{DataApi}
\alias{DataApi}
\title{Data operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataApi
}
\description{
phisWSClientRTools.Data
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_data Get data corresponding to the search parameters given.


get_data_file Get data file


get_data_file_description Get data file description


get_data_file_descriptions_by_search Retrieve data file descriptions corresponding to the search parameters given.


get_data_search Get data corresponding to the search parameters given.


post_data Post data


post_data_file Post data file


post_data_file_path Post data file

}
}

\keyword{datasets}
