% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPostDTO.r
\docType{data}
\name{DataPostDTO}
\alias{DataPostDTO}
\title{DataPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataPostDTO
}
\description{
DataPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{provenanceUri}}{}

\item{\code{objectUri}}{}

\item{\code{variableUri}}{}

\item{\code{date}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
