% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.r
\docType{data}
\name{Dataset}
\alias{Dataset}
\title{Dataset Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Dataset
}
\description{
Dataset Class
}
\section{Fields}{

\describe{
\item{\code{variableURI}}{}

\item{\code{experiment}}{}

\item{\code{provenance}}{}

\item{\code{data}}{}
}}

\keyword{datasets}
