% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetsApi.r
\docType{data}
\name{DatasetsApi}
\alias{DatasetsApi}
\title{Datasets operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DatasetsApi
}
\description{
opensilexWSClient.Datasets
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_data_by_search Get all data corresponding to the search params given


post_dataset_data Post dataset

}
}

\keyword{datasets}
