% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentMetadataDTO.r
\name{DocumentMetadataDTO}
\alias{DocumentMetadataDTO}
\title{DocumentMetadataDTO Class}
\description{
DocumentMetadataDTO Class

DocumentMetadataDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{documentType}}{}

\item{\code{checksum}}{}

\item{\code{creator}}{}

\item{\code{language}}{}

\item{\code{title}}{}

\item{\code{creationDate}}{}

\item{\code{extension}}{}

\item{\code{comment}}{}

\item{\code{concernedItems}}{}

\item{\code{status}}{}

\item{\code{serverFilePath}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DocumentMetadataDTO$new()}}
\item \href{#method-toJSON}{\code{DocumentMetadataDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{DocumentMetadataDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{DocumentMetadataDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{DocumentMetadataDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DocumentMetadataDTO$fromJSONString()}}
\item \href{#method-clone}{\code{DocumentMetadataDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentMetadataDTO$new(
  uri,
  documentType,
  checksum,
  creator,
  language,
  title,
  creationDate,
  extension,
  comment,
  concernedItems,
  status,
  serverFilePath
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentMetadataDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentMetadataDTO$fromJSON(DocumentMetadataDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentMetadataDTO$fromJSONObject(DocumentMetadataDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentMetadataDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentMetadataDTO$fromJSONString(DocumentMetadataDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentMetadataDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
