% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentsApi.r
\docType{data}
\name{DocumentsApi}
\alias{DocumentsApi}
\title{Documents operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DocumentsApi
}
\description{
phisWSClientRTools.Documents
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_document_by_uri Get a document (by receiving it&#39;s uri)


get_documents_metadata_by_search Get all documents metadata corresponding to the searched params given


get_documents_type Get all documents types


post_document_file Post data file


post_documents Save a file


put_document_metadata Update document metadata

}
}

\keyword{datasets}
