% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentMeasureDTO.r
\docType{data}
\name{EnvironmentMeasureDTO}
\alias{EnvironmentMeasureDTO}
\title{EnvironmentMeasureDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EnvironmentMeasureDTO
}
\description{
EnvironmentMeasureDTO Class
}
\section{Fields}{

\describe{
\item{\code{sensorUri}}{}

\item{\code{variableUri}}{}

\item{\code{date}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
