% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironmentMeasurePostDTO.r
\name{EnvironmentMeasurePostDTO}
\alias{EnvironmentMeasurePostDTO}
\title{EnvironmentMeasurePostDTO Class}
\description{
EnvironmentMeasurePostDTO Class

EnvironmentMeasurePostDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sensorUri}}{}

\item{\code{variableUri}}{}

\item{\code{date}}{}

\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EnvironmentMeasurePostDTO$new()}}
\item \href{#method-toJSON}{\code{EnvironmentMeasurePostDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{EnvironmentMeasurePostDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{EnvironmentMeasurePostDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{EnvironmentMeasurePostDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{EnvironmentMeasurePostDTO$fromJSONString()}}
\item \href{#method-clone}{\code{EnvironmentMeasurePostDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasurePostDTO$new(sensorUri, variableUri, date, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasurePostDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasurePostDTO$fromJSON(EnvironmentMeasurePostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasurePostDTO$fromJSONObject(EnvironmentMeasurePostDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasurePostDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasurePostDTO$fromJSONString(EnvironmentMeasurePostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EnvironmentMeasurePostDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
