% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventPostDTO.r
\docType{data}
\name{EventPostDTO}
\alias{EventPostDTO}
\title{EventPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventPostDTO
}
\description{
EventPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{concernedItemsUris}}{}

\item{\code{date}}{}

\item{\code{properties}}{}

\item{\code{description}}{}

\item{\code{creator}}{}
}}

\keyword{datasets}
