% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventPutDTO.r
\docType{data}
\name{EventPutDTO}
\alias{EventPutDTO}
\title{EventPutDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventPutDTO
}
\description{
EventPutDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{concernedItemsUris}}{}

\item{\code{date}}{}

\item{\code{properties}}{}

\item{\code{uri}}{}
}}

\keyword{datasets}
