% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventsApi.r
\name{EventsApi}
\alias{EventsApi}
\title{Events operations}
\description{
phisWSClientRTools.Events
}
\section{Methods}{

\describe{

get_event_annotations Get an event&#39;s annotations


get_event_by_uri Get the event corresponding to the search uri


get_events Get all events corresponding to the search parameters given.


post2 POST event(s)


put1 Update events

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EventsApi$new()}}
\item \href{#method-get_event_annotations}{\code{EventsApi$get_event_annotations()}}
\item \href{#method-get_event_by_uri}{\code{EventsApi$get_event_by_uri()}}
\item \href{#method-get_events}{\code{EventsApi$get_events()}}
\item \href{#method-post2}{\code{EventsApi$post2()}}
\item \href{#method-put1}{\code{EventsApi$put1()}}
\item \href{#method-clone}{\code{EventsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_event_annotations"></a>}}
\subsection{Method \code{get_event_annotations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$get_event_annotations(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_event_by_uri"></a>}}
\subsection{Method \code{get_event_by_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$get_event_by_uri(uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_events"></a>}}
\subsection{Method \code{get_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$get_events(
  page_size,
  page,
  uri,
  type,
  concerned_item_uri,
  concerned_item_label,
  start_date,
  end_date,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post2"></a>}}
\subsection{Method \code{post2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$post2(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put1"></a>}}
\subsection{Method \code{put1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$put1(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
