% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.r
\name{Experiment}
\alias{Experiment}
\title{Experiment Class}
\description{
Experiment Class

Experiment Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{field}}{}

\item{\code{campaign}}{}

\item{\code{place}}{}

\item{\code{alias}}{}

\item{\code{comment}}{}

\item{\code{keywords}}{}

\item{\code{objective}}{}

\item{\code{cropSpecies}}{}

\item{\code{projects}}{}

\item{\code{groups}}{}

\item{\code{contacts}}{}

\item{\code{variables}}{}

\item{\code{sensors}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-toJSON}{\code{Experiment$toJSON()}}
\item \href{#method-fromJSON}{\code{Experiment$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{Experiment$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{Experiment$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Experiment$fromJSONString()}}
\item \href{#method-clone}{\code{Experiment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(
  uri,
  startDate,
  endDate,
  field,
  campaign,
  place,
  alias,
  comment,
  keywords,
  objective,
  cropSpecies,
  projects,
  groups,
  contacts,
  variables,
  sensors
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$fromJSON(ExperimentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$fromJSONObject(ExperimentObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$fromJSONString(ExperimentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
