% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentPostDTO.r
\name{ExperimentPostDTO}
\alias{ExperimentPostDTO}
\title{ExperimentPostDTO Class}
\description{
ExperimentPostDTO Class

ExperimentPostDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{field}}{}

\item{\code{campaign}}{}

\item{\code{place}}{}

\item{\code{alias}}{}

\item{\code{comment}}{}

\item{\code{keywords}}{}

\item{\code{objective}}{}

\item{\code{cropSpecies}}{}

\item{\code{projectsUris}}{}

\item{\code{groupsUris}}{}

\item{\code{contacts}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExperimentPostDTO$new()}}
\item \href{#method-toJSON}{\code{ExperimentPostDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ExperimentPostDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ExperimentPostDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ExperimentPostDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ExperimentPostDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ExperimentPostDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentPostDTO$new(
  startDate,
  endDate,
  field,
  campaign,
  place,
  alias,
  comment,
  keywords,
  objective,
  cropSpecies,
  projectsUris,
  groupsUris,
  contacts
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentPostDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentPostDTO$fromJSON(ExperimentPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentPostDTO$fromJSONObject(ExperimentPostDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentPostDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentPostDTO$fromJSONString(ExperimentPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExperimentPostDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
