% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentsApi.r
\docType{data}
\name{ExperimentsApi}
\alias{ExperimentsApi}
\title{Experiments operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExperimentsApi
}
\description{
phisWSClientRTools.Experiments
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_experiment_data_search Get data corresponding to the search parameters given.


get_experiment_detail Get an experiment


get_experiments_by_search Get all experiments corresponding to the searched params given


post_experiment Post a experiment


put_experiment Update experiment


put_sensors Update the sensors which participates in an experiment


put_variables Update the observed variables of an experiment

}
}

\keyword{datasets}
