% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileDescriptionDTO.r
\docType{data}
\name{FileDescriptionDTO}
\alias{FileDescriptionDTO}
\title{FileDescriptionDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileDescriptionDTO
}
\description{
FileDescriptionDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{date}}{}

\item{\code{concernedItems}}{}

\item{\code{provenanceUri}}{}

\item{\code{metadata}}{}
}}

\keyword{datasets}
