% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileDescriptionWebPathPostDTO.r
\name{FileDescriptionWebPathPostDTO}
\alias{FileDescriptionWebPathPostDTO}
\title{FileDescriptionWebPathPostDTO Class}
\description{
FileDescriptionWebPathPostDTO Class

FileDescriptionWebPathPostDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rdfType}}{}

\item{\code{date}}{}

\item{\code{webPath}}{}

\item{\code{concernedItems}}{}

\item{\code{provenanceUri}}{}

\item{\code{metadata}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FileDescriptionWebPathPostDTO$new()}}
\item \href{#method-toJSON}{\code{FileDescriptionWebPathPostDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{FileDescriptionWebPathPostDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{FileDescriptionWebPathPostDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{FileDescriptionWebPathPostDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{FileDescriptionWebPathPostDTO$fromJSONString()}}
\item \href{#method-clone}{\code{FileDescriptionWebPathPostDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionWebPathPostDTO$new(
  rdfType,
  date,
  webPath,
  concernedItems,
  provenanceUri,
  metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionWebPathPostDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionWebPathPostDTO$fromJSON(FileDescriptionWebPathPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionWebPathPostDTO$fromJSONObject(
  FileDescriptionWebPathPostDTOObject
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionWebPathPostDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionWebPathPostDTO$fromJSONString(FileDescriptionWebPathPostDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileDescriptionWebPathPostDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
