% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialFunding.r
\docType{data}
\name{FinancialFunding}
\alias{FinancialFunding}
\title{FinancialFunding Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FinancialFunding
}
\description{
FinancialFunding Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}
}}

\keyword{datasets}
