% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.r
\docType{data}
\name{Group}
\alias{Group}
\title{Group Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Group
}
\description{
Group Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{level}}{}

\item{\code{description}}{}

\item{\code{users}}{}
}}

\keyword{datasets}
