% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupApi.r
\name{GroupApi}
\alias{GroupApi}
\title{Group operations}
\description{
phisWSClientRTools.Group
}
\section{Methods}{

\describe{

get_group_by_search Get all groups corresponding to the searched params given


get_group_details Get a group


post_group Post a group


put_group Update groups

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GroupApi$new()}}
\item \href{#method-get_group_by_search}{\code{GroupApi$get_group_by_search()}}
\item \href{#method-get_group_details}{\code{GroupApi$get_group_details()}}
\item \href{#method-post_group}{\code{GroupApi$post_group()}}
\item \href{#method-put_group}{\code{GroupApi$put_group()}}
\item \href{#method-clone}{\code{GroupApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_group_by_search"></a>}}
\subsection{Method \code{get_group_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$get_group_by_search(page_size, page, uri, name, level, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_group_details"></a>}}
\subsection{Method \code{get_group_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$get_group_details(group_uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_group"></a>}}
\subsection{Method \code{post_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$post_group(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_group"></a>}}
\subsection{Method \code{put_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$put_group(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
