% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupDTO.r
\docType{data}
\name{GroupDTO}
\alias{GroupDTO}
\title{GroupDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GroupDTO
}
\description{
GroupDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{level}}{}

\item{\code{description}}{}

\item{\code{usersEmails}}{}
}}

\keyword{datasets}
