% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageMetadataDTO.r
\docType{data}
\name{ImageMetadataDTO}
\alias{ImageMetadataDTO}
\title{ImageMetadataDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ImageMetadataDTO
}
\description{
ImageMetadataDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{concernedItems}}{}

\item{\code{configuration}}{}

\item{\code{fileInfo}}{}
}}

\keyword{datasets}
