% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImagesApi.r
\docType{data}
\name{ImagesApi}
\alias{ImagesApi}
\title{Images operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ImagesApi
}
\description{
opensilexWSClient.Images
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_images_by_search Get all images corresponding to the search params given


post_image_file Post data file


post_images_metadata Save a file

}
}

\keyword{datasets}
