% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infrastructure.r
\docType{data}
\name{Infrastructure}
\alias{Infrastructure}
\title{Infrastructure Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Infrastructure
}
\description{
Infrastructure Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}

\item{\code{properties}}{}

\item{\code{rdfType}}{}

\item{\code{rdfTypeLabel}}{}

\item{\code{parent}}{}
}}

\keyword{datasets}
