% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfrastructuresApi.r
\docType{data}
\name{InfrastructuresApi}
\alias{InfrastructuresApi}
\title{Infrastructures operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InfrastructuresApi
}
\description{
phisWSClientRTools.Infrastructures
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_infrastructure_details Get all infrastructure&#39;s details corresponding to the search uri


get_infrastructures_by_search Get all infrastructures corresponding to the search params given

}
}

\keyword{datasets}
