% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogApi.r
\name{LogApi}
\alias{LogApi}
\title{Log operations}
\description{
phisWSClientRTools.Log
}
\section{Methods}{

\describe{

get_data_query_log_search Get data query logs

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogApi$new()}}
\item \href{#method-get_data_query_log_search}{\code{LogApi$get_data_query_log_search()}}
\item \href{#method-clone}{\code{LogApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_query_log_search"></a>}}
\subsection{Method \code{get_data_query_log_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogApi$get_data_query_log_search(
  page_size,
  page,
  user_uri,
  start_date,
  end_date,
  date_sort_asc,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
