% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogoutDTO.r
\docType{data}
\name{LogoutDTO}
\alias{LogoutDTO}
\title{LogoutDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LogoutDTO
}
\description{
LogoutDTO Class
}
\section{Fields}{

\describe{
\item{\code{access_token}}{}
}}

\keyword{datasets}
