% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataFileUAVDTO.r
\name{MetadataFileUAVDTO}
\alias{MetadataFileUAVDTO}
\title{MetadataFileUAVDTO Class}
\description{
MetadataFileUAVDTO Class

MetadataFileUAVDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{installation}}{}

\item{\code{groupPlotType}}{}

\item{\code{groupPlotAlias}}{}

\item{\code{groupPlotUri}}{}

\item{\code{groupPlotSpecies}}{}

\item{\code{pilot}}{}

\item{\code{cameraType}}{}

\item{\code{cameraAlias}}{}

\item{\code{cameraUri}}{}

\item{\code{vectorType}}{}

\item{\code{vectorAlias}}{}

\item{\code{vectorUri}}{}

\item{\code{radiometricTargetAlias}}{}

\item{\code{radiometricTargetUri}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetadataFileUAVDTO$new()}}
\item \href{#method-toJSON}{\code{MetadataFileUAVDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{MetadataFileUAVDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{MetadataFileUAVDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{MetadataFileUAVDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MetadataFileUAVDTO$fromJSONString()}}
\item \href{#method-clone}{\code{MetadataFileUAVDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataFileUAVDTO$new(
  installation,
  groupPlotType,
  groupPlotAlias,
  groupPlotUri,
  groupPlotSpecies,
  pilot,
  cameraType,
  cameraAlias,
  cameraUri,
  vectorType,
  vectorAlias,
  vectorUri,
  radiometricTargetAlias,
  radiometricTargetUri
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataFileUAVDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataFileUAVDTO$fromJSON(MetadataFileUAVDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataFileUAVDTO$fromJSONObject(MetadataFileUAVDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataFileUAVDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataFileUAVDTO$fromJSONString(MetadataFileUAVDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataFileUAVDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
