% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.r
\docType{data}
\name{Method}
\alias{Method}
\title{Method Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Method
}
\description{
Method Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
