% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsApi.r
\docType{data}
\name{MethodsApi}
\alias{MethodsApi}
\title{Methods operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MethodsApi
}
\description{
opensilexWSClient.Methods
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_method_details Get a method


get_methods_by_search Get all methods corresponding to the searched params given


post_method Post method(s)


put_method Update method

}
}

\keyword{datasets}
