% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectPostDTO.r
\docType{data}
\name{ProjectPostDTO}
\alias{ProjectPostDTO}
\title{ProjectPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectPostDTO
}
\description{
ProjectPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{relatedProjects}}{}

\item{\code{financialFunding}}{}

\item{\code{financialReference}}{}

\item{\code{description}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{keywords}}{}

\item{\code{homePage}}{}

\item{\code{administrativeContacts}}{}

\item{\code{coordinators}}{}

\item{\code{scientificContacts}}{}

\item{\code{objective}}{}
}}

\keyword{datasets}
