% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectPutDTO.r
\name{ProjectPutDTO}
\alias{ProjectPutDTO}
\title{ProjectPutDTO Class}
\description{
ProjectPutDTO Class

ProjectPutDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{name}}{}

\item{\code{shortname}}{}

\item{\code{relatedProjects}}{}

\item{\code{financialFunding}}{}

\item{\code{financialReference}}{}

\item{\code{description}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{keywords}}{}

\item{\code{homePage}}{}

\item{\code{administrativeContacts}}{}

\item{\code{coordinators}}{}

\item{\code{scientificContacts}}{}

\item{\code{objective}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectPutDTO$new()}}
\item \href{#method-toJSON}{\code{ProjectPutDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{ProjectPutDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{ProjectPutDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{ProjectPutDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProjectPutDTO$fromJSONString()}}
\item \href{#method-clone}{\code{ProjectPutDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPutDTO$new(
  uri,
  name,
  shortname,
  relatedProjects,
  financialFunding,
  financialReference,
  description,
  startDate,
  endDate,
  keywords,
  homePage,
  administrativeContacts,
  coordinators,
  scientificContacts,
  objective
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPutDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPutDTO$fromJSON(ProjectPutDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPutDTO$fromJSONObject(ProjectPutDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPutDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPutDTO$fromJSONString(ProjectPutDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectPutDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
