% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectsApi.r
\docType{data}
\name{ProjectsApi}
\alias{ProjectsApi}
\title{Projects operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectsApi
}
\description{
phisWSClientRTools.Projects
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_by_id Get a project by its URI.


get_by_search Get all projects corresponding to the searched params given


post3 Post a project


put2 Put project(s)

}
}

\keyword{datasets}
