% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Property.r
\docType{data}
\name{Property}
\alias{Property}
\title{Property Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Property
}
\description{
Property Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{rdfTypeLabels}}{}

\item{\code{relation}}{}

\item{\code{relationLabels}}{}

\item{\code{value}}{}

\item{\code{valueLabels}}{}

\item{\code{domain}}{}

\item{\code{labels}}{}
}}

\keyword{datasets}
