% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyPostDTO.r
\docType{data}
\name{PropertyPostDTO}
\alias{PropertyPostDTO}
\title{PropertyPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PropertyPostDTO
}
\description{
PropertyPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{relation}}{}

\item{\code{value}}{}
}}

\keyword{datasets}
