% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Provenance.r
\docType{data}
\name{Provenance}
\alias{Provenance}
\title{Provenance Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Provenance
}
\description{
Provenance Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{creationDate}}{}

\item{\code{wasGeneratedBy}}{}

\item{\code{documentsUris}}{}
}}

\keyword{datasets}
