% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProvenancesApi.r
\docType{data}
\name{ProvenancesApi}
\alias{ProvenancesApi}
\title{Provenances operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProvenancesApi
}
\description{
phisWSClientRTools.Provenances
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_provenances Get all provenances corresponding to the search params given


post4 Post provenance(s)


put3 Update provenance

}
}

\keyword{datasets}
