% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiometricTargetDTO.r
\name{RadiometricTargetDTO}
\alias{RadiometricTargetDTO}
\title{RadiometricTargetDTO Class}
\description{
RadiometricTargetDTO Class

RadiometricTargetDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{properties}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RadiometricTargetDTO$new()}}
\item \href{#method-toJSON}{\code{RadiometricTargetDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{RadiometricTargetDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{RadiometricTargetDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{RadiometricTargetDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{RadiometricTargetDTO$fromJSONString()}}
\item \href{#method-clone}{\code{RadiometricTargetDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetDTO$new(uri, label, properties)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetDTO$fromJSON(RadiometricTargetDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetDTO$fromJSONObject(RadiometricTargetDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetDTO$fromJSONString(RadiometricTargetDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RadiometricTargetDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
