% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiometricTargetPostDTO.r
\docType{data}
\name{RadiometricTargetPostDTO}
\alias{RadiometricTargetPostDTO}
\title{RadiometricTargetPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RadiometricTargetPostDTO
}
\description{
RadiometricTargetPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{label}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
