% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadiometricTargetsApi.r
\docType{data}
\name{RadiometricTargetsApi}
\alias{RadiometricTargetsApi}
\title{RadiometricTargets operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RadiometricTargetsApi
}
\description{
opensilexWSClient.RadiometricTargets
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_radiometric_targets_by_search Get all radiometric targets corresponding to the search params given


get_radiometric_targets_details Get all radiometric target&#39;s details corresponding to the search uri


post_radiometric_targets Post radiometric(s) target(s) 


put4 Update radiometric targets

}
}

\keyword{datasets}
