% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdfResourceDefinitionDTO.r
\docType{data}
\name{RdfResourceDefinitionDTO}
\alias{RdfResourceDefinitionDTO}
\title{RdfResourceDefinitionDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RdfResourceDefinitionDTO
}
\description{
RdfResourceDefinitionDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
