% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RdfResourceDTO.r
\docType{data}
\name{RdfResourceDTO}
\alias{RdfResourceDTO}
\title{RdfResourceDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RdfResourceDTO
}
\description{
RdfResourceDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}
}}

\keyword{datasets}
