% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectDTO.r
\docType{data}
\name{ScientificObjectDTO}
\alias{ScientificObjectDTO}
\title{ScientificObjectDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectDTO
}
\description{
ScientificObjectDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{geometry}}{}

\item{\code{experiment}}{}

\item{\code{isPartOf}}{}

\item{\code{label}}{}

\item{\code{properties}}{}

\item{\code{uriExperiment}}{}
}}

\keyword{datasets}
