% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectPostDTO.r
\docType{data}
\name{ScientificObjectPostDTO}
\alias{ScientificObjectPostDTO}
\title{ScientificObjectPostDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectPostDTO
}
\description{
ScientificObjectPostDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{geometry}}{}

\item{\code{experiment}}{}

\item{\code{isPartOf}}{}

\item{\code{year}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
