% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectPutDTO.r
\docType{data}
\name{ScientificObjectPutDTO}
\alias{ScientificObjectPutDTO}
\title{ScientificObjectPutDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScientificObjectPutDTO
}
\description{
ScientificObjectPutDTO Class
}
\section{Fields}{

\describe{
\item{\code{rdfType}}{}

\item{\code{geometry}}{}

\item{\code{isPartOf}}{}

\item{\code{label}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
