% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificObjectsApi.r
\name{ScientificObjectsApi}
\alias{ScientificObjectsApi}
\title{ScientificObjects operations}
\description{
phisWSClientRTools.ScientificObjects
}
\section{Methods}{

\describe{

get_scientific_objects_by_search Get all scientific objects corresponding to the searched params given


post_scientific_object Post scientific object(s)


put5 Put scientific object(s) in the given experiment

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScientificObjectsApi$new()}}
\item \href{#method-get_scientific_objects_by_search}{\code{ScientificObjectsApi$get_scientific_objects_by_search()}}
\item \href{#method-post_scientific_object}{\code{ScientificObjectsApi$post_scientific_object()}}
\item \href{#method-put5}{\code{ScientificObjectsApi$put5()}}
\item \href{#method-clone}{\code{ScientificObjectsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_scientific_objects_by_search"></a>}}
\subsection{Method \code{get_scientific_objects_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$get_scientific_objects_by_search(
  page_size,
  page,
  uri,
  experiment,
  alias,
  rdf_type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_scientific_object"></a>}}
\subsection{Method \code{post_scientific_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$post_scientific_object(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put5"></a>}}
\subsection{Method \code{put5()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$put5(uri, experiment, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScientificObjectsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
