% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensor.r
\docType{data}
\name{Sensor}
\alias{Sensor}
\title{Sensor Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Sensor
}
\description{
Sensor Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{model}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{dateOfLastCalibration}}{}

\item{\code{personInCharge}}{}

\item{\code{variables}}{}
}}

\keyword{datasets}
