% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorDTO.r
\docType{data}
\name{SensorDTO}
\alias{SensorDTO}
\title{SensorDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SensorDTO
}
\description{
SensorDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{model}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{dateOfLastCalibration}}{}

\item{\code{personInCharge}}{}
}}

\keyword{datasets}
