% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorsApi.r
\name{SensorsApi}
\alias{SensorsApi}
\title{Sensors operations}
\description{
phisWSClientRTools.Sensors
}
\section{Methods}{

\describe{

get_sensor_data Get data mesured by a sensor


get_sensor_details Get a sensor


get_sensor_profile Get a sensor profile


get_sensors_by_search Get all sensors corresponding to the search params given


post5 Post a sensor


post_profiles Post sensor(s) profile(s)


put6 Update sensors


put_measured_variables1 Update the measured variables of a sensor

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SensorsApi$new()}}
\item \href{#method-get_sensor_data}{\code{SensorsApi$get_sensor_data()}}
\item \href{#method-get_sensor_details}{\code{SensorsApi$get_sensor_details()}}
\item \href{#method-get_sensor_profile}{\code{SensorsApi$get_sensor_profile()}}
\item \href{#method-get_sensors_by_search}{\code{SensorsApi$get_sensors_by_search()}}
\item \href{#method-post5}{\code{SensorsApi$post5()}}
\item \href{#method-post_profiles}{\code{SensorsApi$post_profiles()}}
\item \href{#method-put6}{\code{SensorsApi$put6()}}
\item \href{#method-put_measured_variables1}{\code{SensorsApi$put_measured_variables1()}}
\item \href{#method-clone}{\code{SensorsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sensor_data"></a>}}
\subsection{Method \code{get_sensor_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$get_sensor_data(
  uri,
  variable,
  provenance,
  start_date,
  end_date,
  object,
  page_size,
  page,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sensor_details"></a>}}
\subsection{Method \code{get_sensor_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$get_sensor_details(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sensor_profile"></a>}}
\subsection{Method \code{get_sensor_profile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$get_sensor_profile(uri, page_size, page, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sensors_by_search"></a>}}
\subsection{Method \code{get_sensors_by_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$get_sensors_by_search(
  page_size,
  page,
  uri,
  rdf_type,
  label,
  brand,
  serial_number,
  model,
  in_service_date,
  date_of_purchase,
  date_of_last_calibration,
  person_in_charge,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post5"></a>}}
\subsection{Method \code{post5()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$post5(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_profiles"></a>}}
\subsection{Method \code{post_profiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$post_profiles(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put6"></a>}}
\subsection{Method \code{put6()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$put6(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put_measured_variables1"></a>}}
\subsection{Method \code{put_measured_variables1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$put_measured_variables1(uri, body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SensorsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
