% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShootingConfiguration.r
\docType{data}
\name{ShootingConfiguration}
\alias{ShootingConfiguration}
\title{ShootingConfiguration Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShootingConfiguration
}
\description{
ShootingConfiguration Class
}
\section{Fields}{

\describe{
\item{\code{date}}{}

\item{\code{timestamp}}{}

\item{\code{position}}{}

\item{\code{sensor}}{}
}}

\keyword{datasets}
