% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShootingConfigurationDTO.r
\docType{data}
\name{ShootingConfigurationDTO}
\alias{ShootingConfigurationDTO}
\title{ShootingConfigurationDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShootingConfigurationDTO
}
\description{
ShootingConfigurationDTO Class
}
\section{Fields}{

\describe{
\item{\code{date}}{}

\item{\code{position}}{}

\item{\code{sensor}}{}
}}

\keyword{datasets}
