% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatusException.r
\docType{data}
\name{StatusException}
\alias{StatusException}
\title{StatusException Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StatusException
}
\description{
StatusException Class
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{href}}{}

\item{\code{details}}{}
}}

\keyword{datasets}
