% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyDetails.r
\name{StudyDetails}
\alias{StudyDetails}
\title{StudyDetails Class}
\description{
StudyDetails Class

StudyDetails Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{studyDbId}}{}

\item{\code{studyName}}{}

\item{\code{studyTypeDbId}}{}

\item{\code{studyTypeName}}{}

\item{\code{studyDescription}}{}

\item{\code{seasons}}{}

\item{\code{commonCropName}}{}

\item{\code{trialDbId}}{}

\item{\code{trialName}}{}

\item{\code{startDate}}{}

\item{\code{endDate}}{}

\item{\code{active}}{}

\item{\code{license}}{}

\item{\code{location}}{}

\item{\code{contacts}}{}

\item{\code{dataLinks}}{}

\item{\code{lastUpdate}}{}

\item{\code{additionalInfo}}{}

\item{\code{documentationURL}}{}

\item{\code{studyType}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StudyDetails$new()}}
\item \href{#method-toJSON}{\code{StudyDetails$toJSON()}}
\item \href{#method-fromJSON}{\code{StudyDetails$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{StudyDetails$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{StudyDetails$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StudyDetails$fromJSONString()}}
\item \href{#method-clone}{\code{StudyDetails$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetails$new(
  studyDbId,
  studyName,
  studyTypeDbId,
  studyTypeName,
  studyDescription,
  seasons,
  commonCropName,
  trialDbId,
  trialName,
  startDate,
  endDate,
  active,
  license,
  location,
  contacts,
  dataLinks,
  lastUpdate,
  additionalInfo,
  documentationURL,
  studyType
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetails$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetails$fromJSON(StudyDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetails$fromJSONObject(StudyDetailsObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetails$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetails$fromJSONString(StudyDetailsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyDetails$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
