% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySearchDTO.r
\docType{data}
\name{StudySearchDTO}
\alias{StudySearchDTO}
\title{StudySearchDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StudySearchDTO
}
\description{
StudySearchDTO Class
}
\section{Fields}{

\describe{
\item{\code{commonCropName}}{}

\item{\code{seasonDbId}}{}

\item{\code{studyDbIds}}{}

\item{\code{studyNames}}{}

\item{\code{sortBy}}{}

\item{\code{sortOrder}}{}

\item{\code{page}}{}

\item{\code{pageSize}}{}
}}

\keyword{datasets}
