% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenDTO.r
\docType{data}
\name{TokenDTO}
\alias{TokenDTO}
\title{TokenDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TokenDTO
}
\description{
TokenDTO Class
}
\section{Fields}{

\describe{
\item{\code{grant_type}}{}

\item{\code{username}}{}

\item{\code{password}}{}

\item{\code{client_id}}{}
}}

\keyword{datasets}
