% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenResponseStructure.r
\docType{data}
\name{TokenResponseStructure}
\alias{TokenResponseStructure}
\title{TokenResponseStructure Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TokenResponseStructure
}
\description{
TokenResponseStructure Class
}
\keyword{datasets}
