% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitsApi.r
\docType{data}
\name{TraitsApi}
\alias{TraitsApi}
\title{Traits operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TraitsApi
}
\description{
phisWSClientRTools.Traits
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_trait_details Get a trait


get_traits_by_search Get all Traits corresponding to the searched params given


post_trait Post trait(s)


put_trait Update trait

}
}

\keyword{datasets}
