% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TripletDTO.r
\docType{data}
\name{TripletDTO}
\alias{TripletDTO}
\title{TripletDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TripletDTO
}
\description{
TripletDTO Class
}
\section{Fields}{

\describe{
\item{\code{s}}{}

\item{\code{p}}{}

\item{\code{o}}{}

\item{\code{o_type}}{}

\item{\code{o_lang}}{}

\item{\code{g}}{}
}}

\keyword{datasets}
