% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TripletsApi.r
\docType{data}
\name{TripletsApi}
\alias{TripletsApi}
\title{Triplets operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TripletsApi
}
\description{
opensilexWSClient.Triplets
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

post_triplets Post triplets

}
}

\keyword{datasets}
