% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitsApi.r
\docType{data}
\name{UnitsApi}
\alias{UnitsApi}
\title{Units operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnitsApi
}
\description{
phisWSClientRTools.Units
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_unit_details Get a unit


get_units_by_search Get all units corresponding to the searched params given


post_unit Post unit(s)


put_unit Update unit

}
}

\keyword{datasets}
