% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UriApi.r
\docType{data}
\name{UriApi}
\alias{UriApi}
\title{Uri operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UriApi
}
\description{
opensilexWSClient.Uri
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_ancestors Get all the ancestor of an uri


get_descendants Get all the descendants of an uri


get_instances_by_concept Get all the instances of a concept


get_sibblings Get all the siblings of an Uri


get_type_if_uri_exist get the type of an uri if it exist


get_uri_metadata Get all uri informations


get_uris_by_label get all uri with a given label


is_uri_existing check if an uri exists or not (in the triplestore)

}
}

\keyword{datasets}
