% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserApi.r
\docType{data}
\name{UserApi}
\alias{UserApi}
\title{User operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserApi
}
\description{
phisWSClientRTools.User
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_user_by_search Get all users corresponding to the searched params given


get_user_details Get a user


post_user Post a user


put_user Update users

}
}

\keyword{datasets}
