% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserDTO.r
\docType{data}
\name{UserDTO}
\alias{UserDTO}
\title{UserDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UserDTO
}
\description{
UserDTO Class
}
\section{Fields}{

\describe{
\item{\code{email}}{}

\item{\code{password}}{}

\item{\code{firstName}}{}

\item{\code{familyName}}{}

\item{\code{address}}{}

\item{\code{phone}}{}

\item{\code{affiliation}}{}

\item{\code{orcid}}{}

\item{\code{admin}}{}

\item{\code{groupsUris}}{}
}}

\keyword{datasets}
