% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable.r
\docType{data}
\name{Variable}
\alias{Variable}
\title{Variable Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Variable
}
\description{
Variable Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{label}}{}

\item{\code{comment}}{}

\item{\code{ontologiesReferences}}{}

\item{\code{properties}}{}

\item{\code{trait}}{}

\item{\code{method}}{}

\item{\code{unit}}{}
}}

\keyword{datasets}
