% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesApi.r
\docType{data}
\name{VariablesApi}
\alias{VariablesApi}
\title{Variables operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariablesApi
}
\description{
opensilexWSClient.Variables
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_variable_detail Get a variable


get_variables_by_search Get all variables corresponding to the searched params given


get_variables_details_by_search Get all variables details corresponding to the searched params given


post_variable Post variable(s)


put_variable Update variable

}
}

\keyword{datasets}
