% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDTO.r
\docType{data}
\name{VectorDTO}
\alias{VectorDTO}
\title{VectorDTO Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VectorDTO
}
\description{
VectorDTO Class
}
\section{Fields}{

\describe{
\item{\code{uri}}{}

\item{\code{rdfType}}{}

\item{\code{label}}{}

\item{\code{brand}}{}

\item{\code{serialNumber}}{}

\item{\code{inServiceDate}}{}

\item{\code{dateOfPurchase}}{}

\item{\code{personInCharge}}{}
}}

\keyword{datasets}
