% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabulariesApi.r
\docType{data}
\name{VocabulariesApi}
\alias{VocabulariesApi}
\title{Vocabularies operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VocabulariesApi
}
\description{
phisWSClientRTools.Vocabularies
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_contacts Get all contact properties that can be added to a given rdfType


get_device_properties Get all device properties that can be added to a given rdfType


get_namespaces Get all triplestore namespaces


get_rdfs Get all rdfs properties that can be added to an instance

}
}

\keyword{datasets}
