% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WasGeneratedByDTO.r
\name{WasGeneratedByDTO}
\alias{WasGeneratedByDTO}
\title{WasGeneratedByDTO Class}
\description{
WasGeneratedByDTO Class

WasGeneratedByDTO Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wasGeneratedByDocument}}{}

\item{\code{wasGeneratedByDescription}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WasGeneratedByDTO$new()}}
\item \href{#method-toJSON}{\code{WasGeneratedByDTO$toJSON()}}
\item \href{#method-fromJSON}{\code{WasGeneratedByDTO$fromJSON()}}
\item \href{#method-fromJSONObject}{\code{WasGeneratedByDTO$fromJSONObject()}}
\item \href{#method-toJSONString}{\code{WasGeneratedByDTO$toJSONString()}}
\item \href{#method-fromJSONString}{\code{WasGeneratedByDTO$fromJSONString()}}
\item \href{#method-clone}{\code{WasGeneratedByDTO$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WasGeneratedByDTO$new(wasGeneratedByDocument, wasGeneratedByDescription)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WasGeneratedByDTO$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WasGeneratedByDTO$fromJSON(WasGeneratedByDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONObject"></a>}}
\subsection{Method \code{fromJSONObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WasGeneratedByDTO$fromJSONObject(WasGeneratedByDTOObject)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WasGeneratedByDTO$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WasGeneratedByDTO$fromJSONString(WasGeneratedByDTOJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WasGeneratedByDTO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
